% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psItemContent.R
\name{alignment}
\alias{alignment}
\title{Wrap LaTeX in alignment command}
\arguments{
\item{alignment}{\code{[character(1)]} giving the alignment of the text.
Must be one of:
\itemize{
\item justified
\item left
\item right or
\item center
}

Defaults to \code{'justified'}.}

\item{tex}{\code{[character(1)]} giving some LaTeX string to wrap.}
}
\value{
\code{[character(1)]} LaTeX string.
}
\description{
Wrap LaTeX in alignment command.
}
\keyword{internal}
