% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize.R
\name{anonymize}
\alias{anonymize}
\title{Anonymize names.}
\usage{
anonymize(real_names, lookup_file)
}
\arguments{
\item{real_names}{A character vector with unique names.}

\item{lookup_file}{A character vector of length one as the file to write the lookup table to, or read the lookup table from.
Should be on a safe, private and encrypted volume.
If \code{lookup_file} does not exist, it is created.
See `notes` for caveats.}
}
\value{
A character vector of fake names, same length as \code{real_names}.
Also writes a lookup table to disk at location \code{lookup_file}, if it does not exist already.
}
\description{
Replace real names with random names to anonymize data.
}
\details{
It is sometimes helpful in Q analyses to be able to refer to people-variables by a unique name, though real names often cannot be used in publications to protect participant's data.
This function replaces real names with random names, and also (as a side-effect) writes a lookup table with fake and real names to disk (as a \code{*.csv} file).
Uses names from \pkg{randomNames}.
}
\note{
\strong{
  Despite its name, this function \emph{does not magically anonymize data}, but merely replaces names with randomly drawn fake names.
  It is your responsibility to protect your participants' data.
  If you are unsure, or do not understand the below caveats, \emph{do not rely on this function}.
  \enumerate{
    \item The lookup table with real and fake names must be stored in a safe place, ideally \emph{encrypted} and \emph{no together with the raw data or results}.
    \item Your data may still be deanonymized if it includes other personal information and/or few participants.
    \item Make sure no `real names` are included in your command history, caches or other R objects and scripts.
  }
}
}
\examples{
anonymize(real_names = c("Hillary", "Barack", "George"),
          lookup_file = system.file("extdata",
                                    "example_name_lookup.csv",
                                    package = "pensieve"))
# system.file call only necessary for example, shipped with pensieve
# just as an example, never store lookup file with raw data
# see `notes` for details

}
\author{
Maximilian Held
}
\keyword{internal}

