% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psItemContent.R
\name{capture_disc_output}
\alias{capture_disc_output}
\title{Write file input, read file output}
\usage{
capture_disc_output(fun)
}
\arguments{
\item{fun}{A function which accepts a file as an input and returns a file name as an output.
Helpful for debugging purposes.}
}
\value{
modified function
}
\description{
Function operator to let functions with disk side effects accept R object as input, and return R object as output.
}
\keyword{internal}
