% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R, R/qItems.R
\name{check}
\alias{check}
\alias{check.qItems}
\title{check: validate S3 classes}
\usage{
check(x)

\method{check}{qItems}(x)
}
\arguments{
\item{x}{class object created by respective \code{make_} function.}
}
\description{
Validate S3 classes from pensieve package
}

