% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_fun_args}
\alias{check_fun_args}
\title{Check if function works with specified arguments}
\usage{
check_fun_args(x, y, ...)
}
\arguments{
\item{x}{A function.}

\item{y}{First argument passed to fun.}

\item{...}{Arguments to be passed to fun.}
}
\description{
A wrapper around tryCatch to be used to avoid assertion duplication
}
