% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_no_file}
\alias{check_no_file}
\title{Check if a function already exists}
\usage{
check_no_file(x)
}
\arguments{
\item{x}{[any]\cr
Object to check.}
}
\description{
Wrapper around checkmate::assert_file().
}
