% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_on_dev_machine}
\alias{check_on_dev_machine}
\title{Check if running on main dev machine}
\usage{
check_on_dev_machine(x = c("max"))
}
\arguments{
\item{x}{Character vector of users.}
}
\description{
This test is useful to skip some tests locally and run some additional scripts
}
