% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{correlate}
\alias{correlate}
\title{Correlate Q Sorts.}
\usage{
correlate(sorts, method = "pearson", use = "pairwise.complete.obs")
}
\arguments{
\item{sorts}{An integer array with item handles as first dimension, people as second dimension, arbitrary dimensions thereafter, and item positions in cells.
Dimensions must be named.}

\item{method}{a character string indicating which correlation
    coefficient (or covariance) is to be computed.  One of
    \code{"pearson"} (default), \code{"kendall"}, or \code{"spearman"}:
    can be abbreviated.}

\item{use}{an optional character string giving a
    method for computing covariances in the presence
    of missing values.  This must be (an abbreviation of) one of the strings
    \code{"everything"}, \code{"all.obs"}, \code{"complete.obs"},
    \code{"na.or.complete"}, or \code{"pairwise.complete.obs"}.}
}
\value{
A numerical matrix with people-variables as rows and columns, correlation coefficients in cells.
Class \code{\link{QCors}}.
}
\description{
Correlate Q Sorts.
}
\examples{
correlate(sorts = civicon_2014$qData$sorts[,,"before"])

}
\seealso{
Other correlation functions: \code{\link{QCors}},
  \code{\link{plot.QCors}}
}
