% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inductive_sort.R
\name{count_cooccur}
\alias{count_cooccur}
\title{Create Co-Occurence Matrices.}
\usage{
count_cooccur(ass)
}
\arguments{
\item{ass}{Named list of logical matrices, one for each participant.
Each logical matrix has items as named rows, category indices as columns and logical values in cells.}
}
\value{
An integer array with items as rows and columns, participants as third dimension and cells as co-occurence counts.
}
\description{
Creates co-occurence matrices from logical q-category assignments.
}
\details{
The diagonal is replaced with the \emph{maximum number of categories} for that person, to standardize the entire table.
}
\seealso{
Other import: \code{\link{import_qcat}}
}
\author{
Maximilian Held
}
