% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{count_distros}
\alias{count_distro}
\alias{count_distros}
\title{Count distribution per people-variable}
\usage{
count_distros(data)
}
\arguments{
\item{data}{An integer matrix with item-cases as rows and people-variables as columns.}
}
\value{
A list of dataframes, one for each people-variable, each with a column for the value (\code{x}) and the frequency (\code{freq})-
}
\description{
Counts the number of times an item-case is assigned to each value, yields Q distribution for each people-variable.
}
\details{
Helper function, rarely to be used independently.
Makes sense only when data are integers.
}
\examples{
count_distros(data = civicon_2014$sorts[,,"before"])
# results are all the same, because study used forced Q distribution
}
\keyword{internal}

