% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psItemContent.R
\name{declare_pandoc_var}
\alias{declare_pandoc_var}
\alias{declare_pandoc_fontsize}
\alias{declare_pandoc_lang}
\alias{declare_pandoc_geometry}
\title{Make pandoc tex variable option}
\usage{
declare_pandoc_var(key, value)

declare_pandoc_fontsize(fontsize_global = "10pt")

declare_pandoc_lang(lang = NULL)

declare_pandoc_geometry(paperwidth = 8.5, paperheight = 5.4,
  top = 0.5, bottom = 0.5, left = 0.5, right = 0.5, units = "cm",
  vcentering = TRUE, hcentering = TRUE)
}
\arguments{
\item{key}{\code{[character(1)]} giving the key, such as \code{"geometry"}.}

\item{value}{\code{[character(1)]} giving the value, such as \code{"margin=1in"}.}

\item{fontsize_global}{\code{[character(1)]} giving the document-wide font size.
Must be one of:
\itemize{
\item \code{'10pt'}
\item \code{'11pt'} or
\item \code{'12pt'}
}

Defaults to \code{'10pt'}.}

\item{lang}{\code{[character(1)]} giving a \href{https://tools.ietf.org/html/bcp47}{valid BCP 47 language code} code, such as \code{en_US}. Used for multilingual typsetting support via \href{https://ctan.org/pkg/babel}{LaTeX's babel package} and others.
Must be one of:
\itemize{
\item \code{'ar-DZ'}
\item \code{'ar-IQ'}
\item \code{'ar-JO'}
\item \code{'ar-LB'}
\item \code{'ar-LY'}
\item \code{'ar-MA'}
\item \code{'ar-MR'}
\item \code{'ar-PS'}
\item \code{'ar-SY'}
\item \code{'ar-TN'}
\item \code{'de-DE'}
\item \code{'de-AT'}
\item \code{'de-CH'}
\item \code{'dsb'}
\item \code{'hsb'}
\item \code{'el-polyton'}
\item \code{'en-AU'}
\item \code{'en-CA'}
\item \code{'en-GB'}
\item \code{'en-NZ'}
\item \code{'en-UK'}
\item \code{'en-US'}
\item \code{'grc-ancient'}
\item \code{'la'}
\item \code{'sl'}
\item \code{'fr-CA'}
\item \code{'pt-BR'}
\item \code{'sr'}
\item \code{'af'}
\item \code{'am'}
\item \code{'ar'}
\item \code{'as'}
\item \code{'ast'}
\item \code{'bg'}
\item \code{'bn'}
\item \code{'bo'}
\item \code{'br'}
\item \code{'ca'}
\item \code{'cy'}
\item \code{'cs'}
\item \code{'cop'}
\item \code{'da'}
\item \code{'dv'}
\item \code{'el'}
\item \code{'en'}
\item \code{'eo'}
\item \code{'es'}
\item \code{'et'}
\item \code{'eu'}
\item \code{'fa'}
\item \code{'fr'}
\item \code{'fur'}
\item \code{'ga'}
\item \code{'gd'}
\item \code{'gez'}
\item \code{'gl'}
\item \code{'he'}
\item \code{'hi'}
\item \code{'hr'}
\item \code{'hu'}
\item \code{'hy'}
\item \code{'ia'}
\item \code{'id'}
\item \code{'is'}
\item \code{'it'}
\item \code{'km'}
\item \code{'kmr'}
\item \code{'kn'}
\item \code{'ko'}
\item \code{'lo'}
\item \code{'lt'}
\item \code{'lv'}
\item \code{'ml'}
\item \code{'mn'}
\item \code{'mr'}
\item \code{'nb'}
\item \code{'nl'}
\item \code{'nn'}
\item \code{'no'}
\item \code{'nqo'}
\item \code{'oc'}
\item \code{'pa'}
\item \code{'pms'}
\item \code{'pt'}
\item \code{'rm'}
\item \code{'ro'}
\item \code{'sa'}
\item \code{'se'}
\item \code{'sk'}
\item \code{'sq'}
\item \code{'sr'}
\item \code{'syr'}
\item \code{'ta'}
\item \code{'te'}
\item \code{'th'}
\item \code{'ti'}
\item \code{'tk'}
\item \code{'tr'}
\item \code{'uk'}
\item \code{'ur'} or
\item \code{'vi'}
}

Defaults to \code{NULL}.}

\item{paperwidth}{\code{[numeric(1)]} giving the width of cards in \code{units}.
For good typographical results, should be as close as possible to the \emph{actual} physical measurements of cards encountered by users.
Defaults to \code{8.5}.}

\item{paperheight}{\code{[numeric(1)]} giving the height of cards in \code{units}.
For good typographical results, should be as close as possible to the \emph{actual} physical measurements of cards encountered by users.
Defaults to \code{5.4}.}

\item{top}{\code{[numeric(1)]} giving the margin in \code{units}.
Defaults to \code{0.5}.}

\item{bottom}{\code{[numeric(1)]} giving the margin in \code{units}.
Defaults to \code{0.5}.}

\item{left}{\code{[numeric(1)]} giving the margin in \code{units}.
Defaults to \code{0.5}.}

\item{right}{\code{[numeric(1)]} giving the margin in \code{units}.
Defaults to \code{0.5}.}

\item{units}{\code{[character(1)]} giving the units for the above dimensions, must be one of:
\itemize{
\item \code{"cm"} for metric system,
\item \code{"in"} for inches.
Defaults to \code{"cm"}.
}}

\item{vcentering}{\code{[logical(1)]} indicating whether content should be vertically centered.
Defaults to \code{TRUE}.}

\item{hcentering}{\code{[logical(1)]} indicating whether content should be horizontally centered.
Defaults to \code{TRUE}.}
}
\value{
\code{[character(1)]} giving pandoc variable option.
}
\description{
This function creates pandoc variable key value pairs for LaTeX preamble.
}
\section{Functions}{
\itemize{
\item \code{declare_pandoc_fontsize}: declare \emph{base} font size

\item \code{declare_pandoc_lang}: declare language

\item \code{declare_pandoc_geometry}: declare options for \href{https://ctan.org/pkg/geometry}{LaTeX geometry package}.
}}

\keyword{interal}
