% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_data.R
\name{diag_distro_max}
\alias{diag_distro_max}
\title{Infer maximum grid of q-sorting grid.}
\usage{
diag_distro_max(data)
}
\arguments{
\item{data}{An integer matrix with item-cases as rows and people-variables as columns.}
}
\value{
A positive integer vector of counts of length equal to cover the range of values, named by values.
}
\description{
Infers the maximum q-sorting grid from observed data.
}
\details{
Helper function, rarely to be used independently.
Makes sense only when data are integers.
}
\examples{
diag_distro_max(data = civicon_2014$sorts[,,"before"])
# true, because study used forced Q distribution
}
\keyword{internal}

