% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_data.R
\name{diag_distros}
\alias{diag_distros}
\alias{count_distro}
\alias{count_distros}
\title{Count distribution of same values per people-variable.}
\usage{
diag_distros(sorts)
}
\arguments{
\item{sorts}{An integer array with item handles as first dimension, people as second dimension, arbitrary dimensions thereafter, and item positions in cells.
Dimensions must be named.}
}
\value{
An integer matrix with people-variables as rows, values as columns and counts in cells.
}
\description{
Counts the number of times an item-case is assigned to each value, yields Q distribution for each people-variable.
}
\examples{
diag_distros(sorts = civicon_2014$qData$sorts[,,"before"])
# results are all the same, because study used forced Q distribution

}
\seealso{
Other distribution helpers: \code{\link{diag_distro_max}},
  \code{\link{diag_inside_grid}}, \code{\link{diag_same}}
}
\keyword{internal}
