% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_data.R
\name{diag_forced}
\alias{diag_forced}
\alias{is_forced}
\title{Forced distribution test}
\usage{
diag_forced(data)
}
\arguments{
\item{dataset}{An integer matrix with item-cases as rows and people-variables as columns.}
}
\value{
A logical vector of length one, \code{TRUE} when distribution is forced, else \code{FALSE}.
}
\description{
Tests whether distribution is forced.
}
\details{
Helper function, rarely to be used independently.
Makes sense only when data are integers.
}
\examples{
diag_forced(data = civicon_2014$sorts[,,"before"])
# true, because study used forced Q distribution
}
\keyword{internal}

