% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_data.R
\name{diag_same}
\alias{diag_same}
\alias{is_same}
\title{Same distribution test.}
\usage{
diag_same(sorts)
}
\arguments{
\item{sorts}{An integer array with item handles as first dimension, people as second dimension, arbitrary dimensions thereafter, and item positions in cells.
Dimensions must be named.}
}
\value{
A logical vector of length one, \code{TRUE} when value counts are the same for all people-variables, else \code{FALSE}.
}
\description{
Tests whether the value counts are the same for every people-variable.
}
\note{
Q data that displays the same value counts for all people-variables is likely to stem from a \emph{forced} distribution, where respondents \emph{had} to fill in all physically available slots in the grid, but this need not be so.
Hypothetically, if unlikely, some people-variables might follow exactly the same distributions, even though they had more slots available.
For more information, see \code{\link[=diag_forced]{diag_forced()}}.
}
\examples{
diag_same(sorts = civicon_2014$qData$sorts[,,"before"])
# true, because study used forced Q distribution

}
\seealso{
Other distribution helpers: \code{\link{diag_distro_max}},
  \code{\link{diag_distros}},
  \code{\link{diag_inside_grid}}
}
\keyword{internal}
