% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_ps.R
\name{export_ps}
\alias{export_ps}
\alias{export_ps.default}
\title{Export S3 objects from pensieve in useful formats.}
\usage{
export_ps(x, dir = ".", overwrite = FALSE, ...)

\method{export_ps}{default}(x, dir, ...)
}
\arguments{
\item{x}{An object with one of the pensieve S3 classes.}

\item{dir}{\code{[character(1)]}
giving the directory where the exported objects are written to.
Must be relative from the working directory.
Defaults to the working directory root.}

\item{overwrite}{\code{[logical(1)]}
indicating whether existing files should be overwritten.
Defaults to \code{FALSE}, in which case the function throws an error if a file exists already.}

\item{...}{further arguments to be passed to methods.}
}
\value{
\code{[character()]}
(invisible) giving the paths to the exported files.
}
\description{
Writes S3 objects from pensieve to the file system in some selected formats.
}
\concept{export functions}
