% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{extract}
\alias{extract}
\title{Dimensionality reduction.}
\usage{
extract(cors, nfactors, fa_type = "pca")
}
\arguments{
\item{cors}{A numerical matrix with correlations.}

\item{nfactors}{Positive, non-zero integer scalar, giving the number of components (or factors) to retain.
Must be smaller than the number of observations (here: items) on which \code{cors} are based.}

\item{fa_type}{String giving the type of factor analysis performed.
Currently only \code{"pca"} for principal components analysis (PCA) is allowed.}
}
\description{
Wraps Principal Components Analysis (PCA) to extract component loadings.
}
\examples{
# this just runs the calculations, but also classes and validates the results
cors <- correlate(sorts = civicon_2014$qData$sorts[,,"before"])  # preliminary step
loas <- extract(cors = cors, nfactors = 3, fa_type = "pca")
}
\seealso{
Other analysis functions: \code{\link{correlate}}
}
