% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{find_max_fontsize}
\alias{find_max_fontsize}
\title{Find largest possible fontsize given all other arguments}
\usage{
find_max_fontsize(fontsizes_global_possible = fontsizes_global, ...)
}
\arguments{
\item{fontsizes_global_possible}{\code{[character()]} giving possible fontsizes_local}

\item{...}{Arguments passed to \code{\link[=md2tex_mem]{md2tex_mem()}}}
}
\value{
\code{[character(1)]} giving largest possible fontsize
}
\description{
Finds largest possible fontsize for some markdown to fit on one PDF page.
}
\keyword{internal}
