% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hbs_alpha}
\alias{hbs_alpha}
\title{Pairwise assessments of items on labor and digital technology}
\format{An array of item x item x persons.}
\source{
\url{http://datascience.phil.fau.de/denkzeug/}
}
\usage{
hbs_alpha
}
\description{
A dataset of 18 participants, assessing all pairs of 11 items.
Participants are experts on industrial relations and technology assembled in March 2019 at the Hans-Böckler Stiftung in Dusseldorf, Germany.
This study was an alpha test of the pairwise methodology, conducted as part of the \href{http://datascience.phil.fau.de/denkzeug/}{denkzeug} research project at the Friedrich-Alexander Universität Erlangen-Nürnberg, Germany.
}
\keyword{datasets}
