% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psSort.R
\name{import_psSort}
\alias{import_psSort}
\title{Import Q sorts}
\usage{
import_psSort(x, grid = NULL, items = NULL, ps_coll = NULL,
  lookup = NULL, ...)
}
\arguments{
\item{x}{An object with one of the pensieve S3 classes.}

\item{grid}{A logical matrix giving the available cells for Q sorting.
Accepts arbitrary dimnames from grid.
If any dimnames are missing, sensible defaults will be set.}

\item{items}{\code{[character()]} giving the \emph{participant-facing} \strong{item content}.
Can be named to provide short, \emph{researcher-facing} \strong{item handles}.
\itemize{
\item if \code{dir_bin} is \code{NULL} (default), \code{items} must be text.
Items can be marked up using \href{https://rmarkdown.rstudio.com/authoring_pandoc_markdown.html}{Pandoc Markdown}.
An additional subclass \code{psItemContentText} is prepended and validated.
\item if \code{dir_bin} is given, \code{items} must be file paths, relative from \code{dir_bin}.
An additional subclass \code{psItemContentBin`` is prepended and validated.}lang\code{and}alignment` are ignored.
}}

\item{ps_coll}{\code{\link{AssertCollection}} ps_coll error collection via \code{\link[checkmate:makeAssertCollection]{checkmate::makeAssertCollection()}}, for internal use.}

\item{lookup}{A lookup table}

\item{...}{further arguments to be passed to methods.}
}
\description{
Import Q sorts.
}
