% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inanimatus.R
\name{inanimatus}
\alias{inanimatus}
\title{HTML widget factory for sorting grids.}
\usage{
inanimatus(grid = matrix(data = c(FALSE, TRUE, TRUE, TRUE, FALSE, TRUE), ncol
  = 2), width = NULL, height = NULL)
}
\arguments{
\item{grid}{A logical matrix giving the available cells for Q sorting.
Accepts arbitrary dimnames from grid.
If any dimnames are missing, sensible defaults will be set.}

\item{width}{Fixed width for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{height}{Fixed height for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}
}
\value{
An htmlwidget.
}
\description{
Creates HTML widgets from sorting grids.
Wraps pensieve's inanimatus.js.
End users should use the convenient print method wrappers.
}
