% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{itemConcourse}
\alias{itemConcourse}
\title{Construct items}
\usage{
itemConcourse(full_items = NULL, languages = c("english"),
  type = "textItem", markup = "plain", img_dir = NULL)
}
\arguments{
\item{full_items}{\itemize{
\item for \emph{monolingual studies}, when only one \code{languages} is given: a named character vector of (participant-facing) \strong{full items}, with names as (researcher-facing) \strong{item handles},
\item for \emph{multilingual studies}, when several \code{languages} are given: a character matrix of (participant-facing) \strong{full items}, with rownames as (researcher-facing) \strong{item handles}.
}

Full items must be unique by language, and can be \code{NA} if not available (not recommended).
Names must be unique and valid R names.}

\item{languages}{a character vector of languages, giving the column names.
Defaults to \code{c("english")} for a single language concourse in english.
Must be unique and valid R names.}

\item{type}{a character string giving the \emph{kind} of item stimuli, must be one of:
\itemize{
\item \code{"textItem"} for textual items, in which case cells in \code{concourse} must be plain text.
An additional subclass \code{"textItem"} is prepended and validated.
\item \code{"imageItem"} for image items, in which case cells in \code{concourse} must be file paths, relative from \code{img_dir}.
Images must be \code{*.png}, \code{*.jpg}, \code{*.jpeg} or \code{*.svg}.
An additional class \code{"imageItem"} is prepended and validated.
Defaults to \code{"text"}.
}}

\item{markup}{a character string giving the markup for \code{"textItem"}s.
Defaults to \code{"plain"} for plain text.
Currently only allows \code{"plain"}.}

\item{img_dir}{a character string giving the directory for \code{"imageItem"}s.
Must be relative path \emph{from the working directory}.
Best constructed with \code{\link[base:file.path]{base::file.path()}}.
Defaults to \code{NULL}, in which case images are expected at the working directory root.}
}
\description{
Stores \emph{all} (researcher-facing) \strong{item handles} and (participant-facing) \strong{full items}.
}
\examples{
# monolingual study, text items
monolingual_text <- itemConcourse(
  full_items = c(live_2_work = "Man lives to work.",
                work_2_live = "Man works to live."),
  languages = c("english"),
  type = "textItem",
  markup = "plain"
)

# multilingual study, text items
multilingual_text <- itemConcourse(
  full_items = matrix(
    data = c(
      "Man lives to work.", "Man lebt, um zu arbeiten.",
      "Man works to live.", "Man arbeitet, um zu leben."
    ),
    nrow = 2,
    ncol = 2,
    dimnames = list(items = c("live_2_work", "work_2_live"))
  ),
  languages = c("english", "ngerman"),
  type = "textItem",
  markup = "plain"
)

# monolingual study, image items
monolingual_image <- itemConcourse(
  full_items = c(peach = "peach.jpg",
                pear = "pear.jpg"),
  languages = c("english"),
  img_dir = file.path(system.file(package = "pensieve"), "extdata", "fruit"),
  # these files ship with pensieve
  type = "imageItem"
)

}
