% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{knit_print.ItemConcourse}
\alias{knit_print.ItemConcourse}
\title{Custom print method for knitr}
\usage{
knit_print.ItemConcourse(x, use_js = NULL, ...)
}
\arguments{
\item{x}{a character matrix with full item wording of class \code{\link{ItemConcourse}}, as created by \code{\link[=ItemConcourse]{ItemConcourse()}}.}

\item{use_js}{A logical flag, indicating whether an interactive, java-script variant of the plot should be returned.
Defaults to `NULL`, in which case the appropriate output is inferred from the runtime environment.}

\item{...}{Arguments passed onto other methods.
Not currently used.}
}
\description{
Provides custom print method for knitr.
Can also be invoked manually to open interactive outputs in RStudio.

Plot S3 class.
}
\seealso{
Other plotting functions: \code{\link{QSorts}},
  \code{\link{correlate}}, \code{\link{extract}}
}
