% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QItems.R
\name{knit_print.QConcourse}
\alias{knit_print.QConcourse}
\title{Custom print method for knitr}
\usage{
knit_print.QConcourse(x, use_js = NULL, ...)
}
\arguments{
\item{x}{a character matrix with full item wording of class \code{QConcourse}, as created by \code{\link{QConcourse}}.}

\item{use_js}{Logical flag, indicating whether an interactive, javascript-based version of the plot should be returned.
Suiteable for HTML output (via \code{\link[rmarkdown]{render}} and friends) and the RStudio IDE.
Defaults to \code{NULL}, in which case the appropriate setting is inferred from the runtime environment.}

\item{...}{other parameters to be passed through to plotting functions.}
}
\description{
Provides custom print method for knitr.
Can also be invoked manually to open interactive outputs in RStudio.
}
