% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_print.R
\name{knit_print_docs}
\alias{knit_print_docs}
\title{Documents knit_print}
\arguments{
\item{inline}{\code{[logical(1)]} indicating whether knitr is called from \emph{inline} (\code{r 1+1}) or from a chunk.
Defaults to \code{FALSE}.}

\item{use_js}{\code{[logical(1)]}, indicating whether an interactive, java-script variant of the printed result should be returned.
Applicable only to HTML-outputs.
Defaults to \code{NULL}, in which case the appropriate output is inferred from the runtime environment.}
}
\description{
This is just a placeholder for documentation.
Roxygen templates did not work so well, because calling them would include \emph{all} arguments, and not just the ones actually used
}
