% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psItemContent.R
\name{lang}
\alias{lang}
\title{Set pandoc language option}
\arguments{
\item{lang}{\code{[character(1)]} giving a language code for \emph{all} items, such as \code{en_US}.
Used for multilingual typsetting support via \href{https://ctan.org/pkg/babel}{LaTeX's babel package}.
Must be one of:
\itemize{
\item \code{NULL} (default), in which case there is no multilingual typesetting support.
\item a \href{https://tools.ietf.org/html/bcp47}{valid BCP 47 language code} supported by pandoc.
See \code{pensieve:::format_latex$avail_opts$langs} for all available languages.
Ignored unless \code{type = "text"}.
}}
}
\description{
Set pandoc language option
}
\keyword{internal}
