% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{make_grid}
\alias{make_grid}
\title{Create Q sorting grid.}
\usage{
make_grid(x_range = c(-5, 5), y_range = c(1, 5), pattern = "chessboard",
  offset = "odd")
}
\arguments{
\item{x_range}{An integer vector of length two, giving the range on the x axis.
Defaults to \code{c(-5, 5)}.}

\item{y_range}{An integer vector of length two, giving the range on the y axis.
Defaults to \code{c(1, 5)}.}

\item{pattern}{A character string, giving the pattern of tesselation to use.
Must be \code{"chessboard"} (default), the only currently supported pattern.}

\item{offset}{A character string, giving the rows to be offset.
Must be \code{"even"}, \code{"odd"} or \code{NULL} (default).
Applies only to \code{"honeycomb"} and \code{"brickwall"} patterns, otherwise ignored.}
}
\description{
Creates a grid for Q sorting.
}
\examples{
# this makes the grid, already assigns class
grid <- make_grid(x_range = c(-5, 5), y_range = c(1,5), pattern = "honeycomb", offset = "odd")
}
