% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pensieve-package.r
\docType{package}
\name{pensieve}
\alias{pensieve}
\alias{pensieve-package}
\title{pensieve.}
\description{
Provides tools for the scientific study of human subjectivity.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.maxheld.de/pensieve}
  \item Report bugs at \url{https://github.com/maxheld83/pensieve/issues}
}

}
\author{
\strong{Maintainer}: Maximilian Held \email{info@maxheld.de} (0000-0002-4703-5388)

}
