% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inductive_sort.R
\name{pensieveAssignments}
\alias{pensieveAssignments}
\title{Construct list of categorical assignments}
\usage{
pensieveAssignments(assignments)
}
\arguments{
\item{assignments}{a named list of matrices, with names as a participant names.
Matrices have identical item-handles as row names, and may have arbitrary (ignored) column names.
Matrices must all be of one data type, either
\itemize{
\item \code{logical} for \emph{nominal}-scaled sorts, where a category applies or does not apply,
\item \code{integer} for \emph{ordinally}-scaled sorts, where a category applies to some item \emph{more or less} than to another other item,
\item \code{numeric} for \emph{interval} or \emph{ratio}-scaled sorts, where a category applies to some item \emph{by some amount more or less} than to another item.
Notice that -- counterintuitively -- \emph{categorically}-scaled sorts are not allowed.
}}
}
\description{
Constructs S3 class for pensieve.
}
\examples{
lisa <- matrix(data = c(TRUE, FALSE, FALSE, TRUE),
                      nrow = 2,
                      dimnames = list(handles = c("foo", "bar")))
peter <- matrix(data = c(TRUE, TRUE, TRUE, FALSE, FALSE, FALSE),
                      nrow = 2,
                      dimnames = list(handles = c("foo", "bar")))
assignments <- list(lisa = lisa, peter = peter)
assignments <- pensieveAssignments(list(lisa = lisa, peter = peter))
}
\seealso{
Other S3 classes from `pensieve`: \code{\link{QItemStrata}},
  \code{\link{QPreSorts}}, \code{\link{QSorts}},
  \code{\link{correlate}}, \code{\link{extract}},
  \code{\link{make_grid}}, \code{\link{pensieveConcourse}},
  \code{\link{pensieveItems}},
  \code{\link{pensievePeople}}, \code{\link{score}}
}
