% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{pensieveConcourse}
\alias{pensieveConcourse}
\title{Construct matrix of \emph{all} (researcher-facing) \strong{item handles} and (participant-facing) \strong{full items}}
\usage{
pensieveConcourse(concourse, languages = c("english"), type = "text",
  markup = "plain", img_dir = NULL)
}
\arguments{
\item{concourse}{\itemize{
\item for \emph{monolingual studies}, when only one \code{languages} is given: a named character vector of (participant-facing) \strong{full items}, with names as (researcher-facing) \strong{item handles},
\item for \emph{multilingual studies}, when several \code{languages} are given: a character matrix of (participant-facing) \strong{full items}, with rownames as (researcher-facing) \strong{item handles}.
}

Full items must be unique by language, and can be \code{NA} if not available (not recommended).
Names must be unique and valid R names.}

\item{languages}{a character vector of languages, in the order of columns in \code{concourse}.
Defaults to \code{c("english")}.
Must be unique and valid R names.}

\item{type}{a character string giving the \emph{kind} of item stimuli, must be one of:
\itemize{
\item \code{"text"} for textual items, in which case cells in \code{concourse} must be text.
An additional subclass \code{"pensieveConcourseText"} is prepended and validated.
\item \code{"image"} for image items, in which case cells in \code{concourse} must be file paths, relative from \code{img_dir}.
Images must be \code{*.png}, \code{*.jpg}, \code{*.jpeg} or \code{*.svg}.
An additional class \code{"pensieveConcourseImage"} is prepended and validated.
Defaults to \code{"text"}.
}}

\item{markup}{a character string giving the markup for \code{type = "text"}s.
Defaults to \code{"plain"} for plain text.
Currently only allows \code{"plain"}.}

\item{img_dir}{a character string giving the directory for \code{type = "image"}s.
Must be relative path \emph{from the working directory}.
Best constructed with \code{\link[base:file.path]{base::file.path()}}.
Defaults to \code{NULL}, in which case images are expected at the working directory root.}
}
\description{
Constructs S3 class for pensieve.
}
\examples{
# monolingual study, text items
monolingual_text <- pensieveConcourse(
  concourse = c(live_2_work = "Man lives to work.",
                work_2_live = "Man works to live."),
  languages = c("english"),
  type = "text",
  markup = "plain"
)

# multilingual study, text items
multilingual_text <- pensieveConcourse(
  concourse = matrix(
    data = c(
      "Man lives to work.", "Man lebt, um zu arbeiten.",
      "Man works to live.", "Man arbeitet, um zu leben."
    ),
    nrow = 2,
    ncol = 2,
    dimnames = list(items = c("live_2_work", "work_2_live"))
  ),
  languages = c("english", "ngerman"),
  type = "text",
  markup = "plain"
)

# monolingual study, image items
monolingual_image <- pensieveConcourse(
 concourse = c(peach = "peach.jpg", pear = "pear.jpg"),
 img_dir = file.path(system.file(package = "pensieve"), "extdata", "fruit"),
 # these files ship with pensieve
 type = "image"
)

}
\seealso{
Other S3 classes from `pensieve`: \code{\link{QItemStrata}},
  \code{\link{QPreSorts}}, \code{\link{QSorts}},
  \code{\link{correlate}}, \code{\link{extract}},
  \code{\link{make_grid}}, \code{\link{pensieveItems}},
  \code{\link{pensievePeople}}, \code{\link{score}}
}
