% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{pensieveItems}
\alias{pensieveItems}
\title{Construct items}
\usage{
pensieveItems(pensieve_concourse)
}
\arguments{
\item{pensieve_concourse}{object returned by \code{\link[=pensieveConcourse]{pensieveConcourse()}}}
}
\description{
Stores all item-related information in a list.
}
