% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_sort.R
\name{pensieveOpenSorts}
\alias{pensieveOpenSorts}
\alias{pensieveOpenSort}
\alias{pensieveOpenSorts}
\title{Construct \emph{single} and \emph{multiple} open sort matrix.}
\usage{
pensieveOpenSort(assignments, descriptions = NULL)

pensieveOpenSorts(open_sorts)
}
\arguments{
\item{assignments}{a matrix with item-handles as row names, arbitrary or empty column names, and open sort value in cells.
Matrix must be either
\itemize{
\item \code{logical} for \emph{nominal}-scaled sort, where an open category applies (\code{TRUE}) or does not apply (\code{FALSE}),
\item \code{integer} for \emph{ordinally}-scaled sort, where an open category applies to some item \emph{more} (\code{2nd} rank) \emph{or less} (\code{3rd} rank) than to another other item,
\item \code{numeric} for \emph{interval} or \emph{ratio}-scaled sort, where an open category applies to some item \emph{by some amount more or less} (say \code{2.4} units) than to another item.
Notice that -- counterintuitively -- \emph{categorically}-scaled open sorts are not allowed.
If columns are named, they must be the same as the names in \code{descriptions}.
Either way, \code{assignments} and \code{descriptions} are always \emph{matched by index only}: the first column from \code{assignments}, must be the first element of \code{description}, and so forth.
}}

\item{descriptions}{a character vector giving the open-ended category description provided by the participant.
Can be named.
Defaults to \code{NULL}, in which case the user-defined categories are unknown (not recommended).}

\item{open_sorts}{named list of matrices created by \code{\link[=pensieveOpenSort]{pensieveOpenSort()}}, one for each participant.
Must all be of equal data type and all have the same rows and rownames.}
}
\description{
Constructs S3 class for pensieve.
}
\section{Functions}{
\itemize{
\item \code{pensieveOpenSort}: Creates \emph{individual} open sort.

\item \code{pensieveOpenSorts}: \emph{Combine} individual open sorts in a list.
}}

\examples{
# Lisas open sort, matching by index
assignments <- matrix(data = c(TRUE, FALSE, FALSE, TRUE),
                      nrow = 2,
                      dimnames = list(handles = c("cat", "dog")))
descriptions <- c("a pet which largely takes care of itself",
                  "is known to have saved humans")
lisa <- pensieveOpenSort(assignments = assignments, descriptions = descriptions)

# Peters open sort, matching by name
assignments <- matrix(data = c(TRUE, TRUE, TRUE, FALSE, FALSE, FALSE),
                      nrow = 2,
                      dimnames = list(handles = c("cat", "dog"),
                                      categories = c("in_homes",
                                                     "quiet",
                                                     "herbivore")))
descriptions <- c(in_homes = "Animal found in peoples homes.",
                  quiet = "Does not make a lot of noise.",
                  herbivore = "Eats plants.")
peter <- pensieveOpenSort(assignments = assignments, descriptions = descriptions)

# Rebeccas open sort, without any descriptions provided
assignments <- matrix(data = c(FALSE, FALSE, TRUE, TRUE),
                      nrow = 2,
                      dimnames = list(handles = c("cat", "dog")))
rebecca <- pensieveOpenSort(assignments = assignments, descriptions = NULL)
# providing no description is possible, but makes interpretation hard, if not meaningless.

# now let's combine the individual sort into a list
open_sorts <- pensieveOpenSorts(open_sorts = list(lisa = lisa, peter = peter, rebecca = rebecca))

}
\seealso{
Other S3 classes from `pensieve`: \code{\link{QItemStrata}},
  \code{\link{QPreSorts}}, \code{\link{QSorts}},
  \code{\link{correlate}}, \code{\link{extract}},
  \code{\link{make_grid}}, \code{\link{pensieveConcourse}},
  \code{\link{pensieveItems}},
  \code{\link{pensievePeople}}, \code{\link{score}}
}
