% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people.R
\name{pensievePeople}
\alias{pensievePeople}
\title{Construct tibble with additional participant information}
\usage{
pensievePeople(people)
}
\arguments{
\item{people}{A dataframe or tibble, with one row per participant.
First column must be (short form of) participant name and valid R name.
Later columns can have arbitrary additional information about participants.
Columns must be named.}
}
\description{
Constructs S3 class for pensieve.
}
\examples{
people <- pensievePeople(data.frame(Name = c("Ann", "Kim", "Joe"),
                                    Gender = c("female", "other", "male"),
                                    stringsAsFactors = FALSE))
}
\seealso{
Other S3 classes from `pensieve`: \code{\link{QItemStrata}},
  \code{\link{QPreSorts}}, \code{\link{QSorts}},
  \code{\link{correlate}}, \code{\link{extract}},
  \code{\link{make_grid}}, \code{\link{pensieveConcourse}},
  \code{\link{pensieveItems}}, \code{\link{score}}
}
