% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people.R
\name{PeopleInfo}
\alias{PeopleInfo}
\title{Check and make PeopleInfo}
\usage{
PeopleInfo(people_info, validate = TRUE)
}
\arguments{
\item{people_info}{A dataframe or tibble, with one row per participant.
First column must be participant name, same as in other objects.}

\item{validate}{a logical flag, indicating whether the object will be validated on construction.
Defaults to `TRUE`.}
}
\description{
Checks and makes PeopleInfo, a tibble with arbitrary additional information on the participating people-variables.

Construct S3 class for pensieve.
}
\examples{
PeopleInfo(data.frame(Name = c("Ann", "Kim", "Joe"),
                          Gender = c("female", "other", "male")))
}
\seealso{
Other S3 classes from `pensieve`: \code{\link{ItemConcourse}},
  \code{\link{QItemStrata}}, \code{\link{QPreSorts}},
  \code{\link{QSorts}}, \code{\link{correlate}},
  \code{\link{extract}}, \code{\link{make_grid}},
  \code{\link{score}}

Other import helpers: \code{\link{QItemStrata}},
  \code{\link{QPreSorts}}, \code{\link{QSorts}},
  \code{\link{anonymize}}
}
