% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{plot.QCors}
\alias{plot.QCors}
\title{Plot Q correlations.}
\usage{
\method{plot}{QCors}(x, use_js = NULL)
}
\arguments{
\item{x}{An object of class \code{\link{QCors}}.}

\item{use_js}{A logical flag, indicating whether an interactive, java-script variant of the plot should be returned.
Defaults to `NULL`, in which case the appropriate output is inferred from the runtime environment.}
}
\description{
Produces a heatmap from a given correlation matrix.
}
\seealso{
Other printing functions: \code{\link{knit_print.QConcourse}}

Other correlation functions: \code{\link{QCors}},
  \code{\link{correlate}}
}
