% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qItems.R
\name{print.QItems}
\alias{print.QItems}
\title{Print Q items}
\usage{
\method{print}{QItems}(x, iprint = NULL, ...)
}
\arguments{
\item{x}{a character matrix with full item wording of class \code{QItems}, as created by \code{\link{QItems}}.}

\item{iprint}{Logical flag, indicating whether an interactive version of the plot should be returned.
Suiteable for HTML output (via \code{\link[rmarkdown]{render}} and friends) and the RStudio IDE.
Defaults to \code{NULL}, in which case the appropriate setting is inferred from the runtime environment.}

\item{...}{other parameters to be passed through to printing functions.}
}
\value{
Prints a properly formatted character string.
}
\description{
Prints Q items in a form convenient for the runtime, or for documents.
}
