% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psClosedSorts.R
\name{psClosedSorts}
\alias{psClosedSorts}
\alias{validate_S3.psClosedSorts}
\alias{plot.psClosedSorts}
\title{Store multiple sorts as a numeric matrix}
\usage{
psClosedSorts(csorts)

\method{validate_S3}{psClosedSorts}(x, ...)

\method{plot}{psClosedSorts}(x, column = 1, use_js = NULL, ...)
}
\arguments{
\item{csorts}{\code{[array()]}
An numeric array with people as rows, item handles as columns, arbitrary dimensions thereafter (such as for multiple sorting dimensions), and item positions in cells.}

\item{x}{An object with one of the pensieve S3 classes.}

\item{...}{Arguments passed onto other methods.
Not currently used.}

\item{column}{Positive integer scalar, giving the column of the psClosedSorts object to plot.
Defaults to \code{1}, in which case the first column is plotted.}

\item{use_js}{A logical flag, indicating whether an interactive, java-script variant of the plot should be returned.
Defaults to \code{NULL}, in which case the appropriate output is inferred from the runtime environment.}
}
\value{
A numeric matrix of class \link{psClosedSorts}.
}
\description{
Canonical storage for closed sorts.

Plot S3 class.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_S3}: Validation

\item \code{plot}: plotting
}}

\examples{
# construction ====
# simple case with one condition of instruction
csorts <- matrix(
  data = c(-1, 0, 1, -1),
  nrow = 2,
  byrow = TRUE,
  dimnames = list(
    people = c("Lisa", "Peter"),
    items = c("live_2_work", "work_2_live")
  )
)
csorts <- psClosedSorts(csorts)

# two conditions of instruction are planes in a 3-d array
csorts_multiple_conds <- array(
  data = c(-1, 0, 1, -1, 0, 1, -1, 0),
  dim = c(2, 2, 2),
  dimnames = list(
    people = c("Lisa", "Peter"),
    items = c("live_2_work", "work_2_live"),
    conditions = c("desirable", "probable")
  )
)
csorts_multiple_conds <- psClosedSorts(csorts_multiple_conds)
}
\seealso{
Other S3 classes from `pensieve`: \code{\link{correlate}},
  \code{\link{extract}}, \code{\link{psGrid}},
  \code{\link{psItemContent}}, \code{\link{psOpenSorts}},
  \code{\link{psOpenSort}}, \code{\link{psPeople}},
  \code{\link{score}}

Other plotting functions: \code{\link{correlate}},
  \code{\link{extract}}
}
\concept{S3 classes from `pensieve`}
\concept{plotting functions}
