% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{psConcourse}
\alias{psConcourse}
\title{Constructs matrix of \emph{all} \strong{full items}.}
\usage{
psConcourse(concourse, type = "text", markup = "plain", babel = TRUE,
  img_dir = NULL)
}
\arguments{
\item{concourse}{A matrix with:
\itemize{
\item researcher-facing \strong{item handles} as row names,
\item \strong{languages} as column names and
\item participant-facing \strong{full items} in cells.
}

For \emph{monolingual studies}, use one column only.

Full items must be unique by language, and can be \code{NA} if not available (not recommended).
Names must be unique and valid R names.}

\item{type}{a character string giving the \emph{kind} of full item stimuli, must be one of:
\itemize{
\item \code{"text"} for textual items, in which case cells in \code{concourse} must be text.
An additional subclass \code{"psConcourseText"} is prepended and validated.
\item \code{"image"} for image items, in which case cells in \code{concourse} must be file paths, relative from \code{img_dir}.
Images must be \code{*.png}, \code{*.jpg}, \code{*.jpeg} or \code{*.svg}.
An additional class \code{"psConcourseImage"} is prepended and validated.
Defaults to \code{"text"}.
}}

\item{markup}{a character string giving the markup for \code{type = "text"}s.
Defaults to \code{"plain"} for plain text.
Currently only allows \code{"plain"}.
Ignored unless \code{type = "text"}.}

\item{babel}{a logical flag, indicating whether \href{https://ctan.org/pkg/babel}{LaTeX's babel package} should be used for multilingual support.
If \code{TRUE} (default), column names in \code{concourse} must be \href{http://ctan.math.washington.edu/tex-archive/macros/latex/required/babel/base/babel.pdf}{valid babel languages}.
Ignored unless \code{type = "text"}.}

\item{img_dir}{a character string giving the directory for \code{type = "image"}s.
Must be relative path \emph{from the working directory}.
Best constructed with \code{\link[base:file.path]{base::file.path()}}.
Defaults to \code{NULL}, in which case images are expected at the working directory root.
Ignored unless \code{type = "image"}.}
}
\description{
Constructs S3 class for pensieve.
}
\examples{
# multilingual study, text items
multilingual_text <- psConcourse(
  concourse = matrix(
    data = c(
      "Man lives to work.", "Man lebt, um zu arbeiten.",
      "Man works to live.", "Man arbeitet, um zu leben."
    ),
    nrow = 2,
    ncol = 2,
    dimnames = list(
      items = c("live_2_work", "work_2_live"),
      languages = c("english", "ngerman"))
  ),
  type = "text",
  markup = "plain",
  babel = TRUE
)

# monolingual study, image items
monolingual_image <- psConcourse(
  concourse = matrix(
    data = c("peach.jpg",
             "pear.jpg"),
    nrow = 2,
    ncol = 1,
    dimnames = list(
      items = c("peach", "pear"),
      languages = c("english")
   )),
 type = "image",
 img_dir = file.path(system.file(package = "pensieve"), "extdata", "fruit")
 # these files ship with pensieve
)

}
\seealso{
Other S3 classes from `pensieve`: \code{\link{QItemStrata}},
  \code{\link{QPreSorts}}, \code{\link{QSorts}},
  \code{\link{correlate}}, \code{\link{extract}},
  \code{\link{make_grid}}, \code{\link{psItems}},
  \code{\link{psOpenSorts}}, \code{\link{psPeople}},
  \code{\link{score}}
}
