% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psConcourseExtra.R
\name{psConcourseExtra}
\alias{psConcourseExtra}
\title{Store additional arbitrary information about psConcourse items}
\usage{
psConcourseExtra(concourse_extra)
}
\arguments{
\item{concourse_extra}{A data frame or tibble with additional information.
There must be as many rows as rows in \code{\link[=psConcourse]{psConcourse()}}, and in the same order.
If there is a first column named \code{items}, it must be the same as the rownames of \code{\link[=psConcourse]{psConcourse()}}.}
}
\value{
A tibble of class \code{psConcourseExtra}
}
\description{
Store additional arbitrary information about psConcourse items
}
