% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/people.R
\name{psPeople}
\alias{psPeople}
\title{Construct tibble with additional participant information}
\usage{
psPeople(people)
}
\arguments{
\item{people}{A dataframe or tibble, with one row per participant.
First column must be (short form of) participant name and valid R name.
Later columns can have arbitrary additional information about participants.
Columns must be named.}
}
\description{
Constructs S3 class for pensieve.
}
\examples{
people <- psPeople(data.frame(Name = c("Ann", "Kim", "Joe"),
                                    Gender = c("female", "other", "male"),
                                    stringsAsFactors = FALSE))
}
\seealso{
Other S3 classes from `pensieve`: \code{\link{QPreSorts}},
  \code{\link{QSorts}}, \code{\link{correlate}},
  \code{\link{extract}}, \code{\link{make_grid}},
  \code{\link{psConcourse}}, \code{\link{psItemContent}},
  \code{\link{psItems}}, \code{\link{psOpenSorts}},
  \code{\link{psOpenSort}}, \code{\link{score}}
}
\concept{S3 classes from `pensieve`}
