% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{QCors}
\alias{QCors}
\title{Check and make Q correlations.}
\usage{
QCors(cors, validate = TRUE)
}
\arguments{
\item{cors}{A numerical matrix with correlations.}

\item{validate}{A logical flag, indicating whether the object will be validated.
Defaults to `TRUE`.}
}
\value{
A numerical matrix with people-variables as rows and columns, correlation coefficients in cells.
Class \code{\link{QCors}}.
}
\description{
Check and make Q correlations.
}
\seealso{
Other validation helpers: \code{\link{QConcourse}},
  \code{\link{QItemStrata}}, \code{\link{QPeopleFeatures}},
  \code{\link{QPreSorts}}, \code{\link{QSorts}}

Other correlation functions: \code{\link{correlate}},
  \code{\link{plot.QCors}}
}
