% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QData.R
\name{QData}
\alias{QData}
\title{Check and make QData.}
\usage{
QData(presorts = NULL, validate = TRUE)
}
\arguments{
\item{presorts}{An integer matrix, with named rows as item handles, named columns as participant names and cells as presorts.
\code{-1L} for `negative`, \code{0L} for `neutral` and \code{1L} for `positive`.}

\item{validate}{A logical flag, indicating whether the object will be validated.
Defaults to \code{TRUE}.}
}
\description{
Checks and makes QData
}
\details{
All data gathered during a Q study.
}
\note{
\code{presorts} are stored as \code{integer()} because R does not allow factor matrices.
Pre-sorting piles are, of course, \emph{categorical} information and should be treated as such.
}
\seealso{
Other import helpers: \code{\link{QItems}},
  \code{\link{anonymize}}, \code{\link{check.QData}}

Other validation helpers: \code{\link{QItems}}
}

