% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QItems.R
\name{QItemStrata}
\alias{QItemStrata}
\title{Check and make QItemStrata}
\usage{
QItemStrata(strata, validate = TRUE)
}
\arguments{
\item{strata}{A logical array of arbitrary dimensions, with first dimension (rows) as item handles, and higher dimensions as orthogonal sampling strata.
Rownames must be a subset of the rownames from \link{QConcourse}.}

\item{validate}{A logical flag, indicating whether the object will be validated.
Defaults to `TRUE`.}
}
\description{
Checks and makes QItemStrata, the item sampling structure
}
\seealso{
Other validation helpers: \code{\link{QConcourse}},
  \code{\link{QCors}}, \code{\link{QPeopleFeatures}},
  \code{\link{QPreSorts}}, \code{\link{QSorts}}

Other import helpers: \code{\link{QConcourse}},
  \code{\link{QPeopleFeatures}}, \code{\link{QPreSorts}},
  \code{\link{QSorts}}, \code{\link{anonymize}},
  \code{\link{check.QPreSorts}}
}
