% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QData.R
\name{QPeopleFeatures}
\alias{QPeopleFeatures}
\title{Check and make QPeopleFeatures}
\usage{
QPeopleFeatures(p_feat, validate = TRUE)
}
\arguments{
\item{p_feat}{A tibble, with one row per participant.
First column must be the participant names, same as the rownames from \code{\link{QSorts}}.}

\item{validate}{A logical flag, indicating whether the object will be validated.
Defaults to `TRUE`.}
}
\description{
Checks and makes QPeopleFeatures, a tibble with arbitrary additional information on the participating people-variables.
}
\seealso{
Other validation helpers: \code{\link{QConcourse}},
  \code{\link{QCors}}, \code{\link{QItemStrata}},
  \code{\link{QPreSorts}}, \code{\link{QSorts}}

Other import helpers: \code{\link{QConcourse}},
  \code{\link{QItemStrata}}, \code{\link{QPreSorts}},
  \code{\link{QSorts}}, \code{\link{anonymize}},
  \code{\link{check.QPreSorts}}
}
