% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psClosedSorts.R
\name{QPreSorts}
\alias{QPreSorts}
\title{Check and make QPreSorts}
\usage{
QPreSorts(presorts, validate = TRUE)
}
\arguments{
\item{presorts}{An integer matrix, with named rows as item handles, named columns as participant names and cells as presorts.
\code{-1L} for \code{negative}, \code{0L} for \code{neutral} and \code{1L} for \code{positive}.}

\item{validate}{a logical flag, indicating whether the object will be validated on construction.
Defaults to \code{TRUE}.}
}
\description{
Checks and makes QPreSorts

Construct S3 class for pensieve.
}
\note{
\code{presorts} are stored as \code{integer()} because R does not allow factor matrices.
Pre-sorting piles are, of course, \code{categorical} information and should be treated as such.
}
\seealso{
Other S3 classes from `pensieve`: \code{\link{QItemStrata}},
  \code{\link{QSorts}}, \code{\link{correlate}},
  \code{\link{extract}}, \code{\link{make_grid}},
  \code{\link{psConcourse}}, \code{\link{psItems}},
  \code{\link{psOpenSorts}}, \code{\link{psOpenSort}},
  \code{\link{psPeople}}, \code{\link{score}}

Other import helpers: \code{\link{QItemStrata}},
  \code{\link{QSorts}}, \code{\link{anonymize}}
}
\concept{S3 classes from `pensieve`}
\concept{import helpers}
