% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorts.R
\name{QSorts}
\alias{QSorts}
\title{Check and make QSorts}
\usage{
QSorts(sorts, validate = TRUE)
}
\arguments{
\item{sorts}{An integer array with item handles as first dimension, people as second dimension, arbitrary dimensions thereafter, and item positions in cells.
Dimensions must be named.}

\item{validate}{A logical flag, indicating whether the object will be validated on construction.
Defaults to `TRUE`.}
}
\description{
Checks and makes QSorts

Construct S3 class.
}
\seealso{
Other S3 classes from `pensieve`: \code{\link{QConcourse}},
  \code{\link{QItemStrata}}, \code{\link{QPeopleFeatures}},
  \code{\link{QPreSorts}}, \code{\link{correlate}},
  \code{\link{extract}}, \code{\link{score}}

Other import helpers: \code{\link{QConcourse}},
  \code{\link{QItemStrata}}, \code{\link{QPeopleFeatures}},
  \code{\link{QPreSorts}}, \code{\link{anonymize}},
  \code{\link{check}}
}
