% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markup2vector.R
\name{render_chain}
\alias{render_chain}
\title{Render a list of markdown vectors to a desired format.}
\usage{
render_chain(l, format, ...)
}
\arguments{
\item{l}{\code{[list()]}
giving \code{x}s to be passed to \code{\link[=md2tex_mem]{md2tex_mem()}}.}

\item{format}{\code{[character(1)]} giving the output format to render items in.

Must be one of:
\itemize{
\item \code{'tex'}
\item \code{'pdf'} (\strong{default})
\item \code{'svg'} or
\item \code{'grob'}
}}

\item{...}{Arguments passed on to \code{md2tex_mem}
\describe{
  \item{x}{\code{[character()]} \emph{or} \code{[raw()]} giving the input.}
  \item{path_in}{\code{[character(1)]} giving path to use for input file \emph{with or without extension}.
Defaults to \code{"foo"}.
Useful for debugging.}
  \item{fontsize_local}{\code{[character(1)]} giving a valid \href{https://en.wikibooks.org/wiki/LaTeX/Fonts#Sizing_text}{LaTeX font size}.

Must be one of:
\itemize{
\item \code{NULL} in which case the default local fontsize is used (\strong{default})
\item \code{'tiny'}
\item \code{'scriptsize'}
\item \code{'footnotesize'}
\item \code{'small'}
\item \code{'normalsize'}
\item \code{'large'}
\item \code{'Large'}
\item \code{'LARGE'}
\item \code{'huge'} or
\item \code{'Huge'}
}}
  \item{alignment}{\code{[character(1)]} giving the alignment of the text.

Must be one of:
\itemize{
\item \code{'justified'} (\strong{default})
\item \code{'left'}
\item \code{'right'} or
\item \code{'center'}
}}
}}
}
\value{
\code{[list()]}
of output format objects.
}
\description{
Goes through the conversion chain as long as necessary to return the desired output.
}
\keyword{internal}
