% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psItemContent.R
\name{render_items}
\alias{render_items}
\title{Render a list of text items to desired format.}
\usage{
render_items(items, format = "tex")
}
\arguments{
\item{items}{\code{[character()]} giving the \emph{participant-facing} \strong{item content}.
Can be named to provide short, \emph{researcher-facing} \strong{item handles}.
\itemize{
\item if \code{dir_bin} is \code{NULL} (default), \code{items} must be text.
Items can be marked up using \href{https://rmarkdown.rstudio.com/authoring_pandoc_markdown.html}{Pandoc Markdown}.
An additional subclass \code{psItemContentText} is prepended and validated.
\item if \code{dir_bin} is given, \code{items} must be file paths, relative from \code{dir_bin}.
An additional subclass \code{psItemContentBin`` is prepended and validated.}lang\code{and}alignment` are ignored.
}}

\item{format}{\code{[character(1)]} giving the output format to render items in.

Must be one of:
\itemize{
\item \code{'tex'}
\item \code{'pdf'} (\strong{default})
\item \code{'svg'} or
\item \code{'grob'}
}}
}
\value{
\code{[list()]}
of vectors in output format (\code{[character()]} or \code{[raw()]}), one list element per item.
}
\description{
Goes through the conversion chain as long as necessary to return the desired output.
}
\keyword{internal}
