% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_accio.R
\name{run_accio}
\alias{run_accio}
\title{Run accio}
\usage{
run_accio()
}
\description{
Run accio, the web frontend for pensieve, if available.
}
\details{
Accio is the closed-source web frontend for pensieve.
Learn more at https://pensieve.maxheld.de.
}
