% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\alias{QScores}
\alias{check.QScores}
\title{Factor scores.}
\usage{
score(loas, sorts)

QScores(scores, validate = TRUE)

\method{check}{QScores}(x)
}
\arguments{
\item{loas}{A numerical matrix with loadings.}

\item{sorts}{An integer array with item handles as first dimension, people as second dimension, arbitrary dimensions thereafter, and item positions in cells.
Dimensions must be named.}

\item{scores}{A numerical matrix with factor scores.}

\item{validate}{A logical flag, indicating whether the object will be validated on construction.
Defaults to `TRUE`.}

\item{x}{class object created by respective constructor function.}
}
\description{
Calculates the factor scores as loadings-weighted averages of raw Q-Sorts (regression scores).

Construct S3 class.

Check S3 class.
}
\section{Methods (by generic)}{
\itemize{
\item \code{check}: validation
}}

\examples{
sorts <- civicon_2014$qData$sorts[,,"before"]  # preparatory step
cors <- correlate(sorts = sorts)  # preparatory step
loas <- extract(cors = cors, nfactors = 3, fa_type = "pca")  # preparatory step
scores <- score(loas = loas, sorts = sorts)
# this just assigns the class, without validation (not recommended)
scores <- QScores(scores = scores, validate = FALSE)
# this validates the class
check(x = scores)
}
\seealso{
Other analysis functions: \code{\link{correlate}},
  \code{\link{extract}}

Other S3 classes from `pensieve`: \code{\link{QItemConcourse}},
  \code{\link{QItemStrata}}, \code{\link{QPeopleFeatures}},
  \code{\link{QPreSorts}}, \code{\link{QSorts}},
  \code{\link{correlate}}, \code{\link{extract}},
  \code{\link{make_grid}}

Other validation functions: \code{\link{QSorts}},
  \code{\link{check}}, \code{\link{correlate}},
  \code{\link{extract}}, \code{\link{make_grid}}
}
