% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{validate_S3}
\alias{validate_S3}
\alias{validate_S3.default}
\alias{check_S3}
\alias{test_S3}
\alias{expect_S3}
\alias{assert_S3}
\alias{need_S3}
\title{Validate S3 classes from pensieve}
\usage{
validate_S3(x, ...)

\method{validate_S3}{default}(x, ps_coll = NULL, ...)

check_S3(x)

test_S3(x)

expect_S3(x, info = NULL, label = NULL)

assert_S3(x, collection = NULL, var.name = NULL)

need_S3(x, label = NULL)
}
\arguments{
\item{x}{An object with one of the pensieve S3 classes.}

\item{...}{further arguments to be passed to methods.}

\item{ps_coll}{\code{\link{AssertCollection}} ps_coll error collection via \code{\link[checkmate:makeAssertCollection]{checkmate::makeAssertCollection()}}, for internal use.}

\item{info}{[\code{character(1)}]\cr
See \code{\link[testthat]{expect_that}}}

\item{label}{[\code{character(1)}]\cr
See \code{\link[testthat]{expect_that}}}

\item{collection}{[\code{\link{AssertCollection}}]\cr
If an \code{\link{AssertCollection}} is provided, the error message is stored
in it. If \code{NULL}, an exception is raised if \code{res} is not
\code{TRUE}.}

\item{var.name}{[\code{character(1)}]\cr
The custom name for \code{x} as passed to any \code{assert*} function.
Defaults to a heuristic name lookup.}
}
\description{
Use \code{validate_S3()}, \code{check_S3()}, \code{test_S3()}, \code{assert_S3()}, \code{expect_S3()} and \code{need_S3()} to validate S3 objects from this package.
}
\details{
All S3 classes in pensieve have \code{validate_S3()} methods, which return \code{NULL} on success, or a character vector of arbitrary length with validation failures.
Five alternative generics use the same underlying \code{\link[=validate_S3]{validate_S3()}} methods, but differ in their returns:
\itemize{
\item \strong{\href{https://github.com/mllg/checkmate}{checkmate}} package extensions:
\itemize{
\item \code{check_S3()} returns \code{TRUE} or the error message as a character string,
\item \code{assert_S3()} returns \code{x} invisibly or throws an error,
\item \code{test_S3()} returns \code{TRUE} or \code{FALSE},
}
\item \strong{\href{https://github.com/hadley/testthat}{testthat})} package extension:
\itemize{
\item \code{expect_S3()} always returns an \code{\link[testthat:expectation]{testthat::expectation()}} for internal use in testing,
}
\item \strong{\href{https://shiny.rstudio.com}{shiny}} package extension:
\itemize{
\item \code{need_S3()} returns \code{NULL} or the error message for internal use with the accio web frontend inside \code{\link[shiny:validate]{shiny::validate()}}.
}
}
}
\examples{
# just for testing; never build objects like this by hand
good_obj <- structure(
  .Data = "I am an item.",
  class = c("psItemContentText", "psItemContent", "character")
)
bad_obj <- structure(
  .Data = 1L, # must be character
  class = c("psItemContentText", "psItemContent", "character")
)

validate_S3(good_obj)
validate_S3(bad_obj)
check_S3(good_obj)
check_S3(bad_obj)
test_S3(good_obj)
test_S3(bad_obj)
expect_S3(good_obj)
# expect_S3(bad_obj) # this errors out
assert_S3(good_obj)
# assert_S3(bad_obj)  # this errors out
need_S3(good_obj)
need_S3(bad_obj)


}
