% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymize.R
\name{anonymize}
\alias{anonymize}
\title{Anonymize names.}
\usage{
anonymize(real_names, lookup_file)
}
\arguments{
\item{real_names}{A character vector with unique names.}

\item{lookup_file}{A character vector of length one as the file to write the lookup table to, or read the lookup table from.
See `details` for more, `notes` for caveats.}
}
\value{
A character vector of fake names, same length as \code{real_names}.
Also writes a lookup table to disk at location \code{lookup_file}, if it does not exist already.
}
\description{
Replace real names with random names to anonymize data.
}
\details{
It is sometimes helpful in Q analyses to be able to refer to people-variables by a unique name, though real names often cannot be used in publications to protect participant's data.

This function looks up \code{real_names} in the \code{lookup_file} and returns the respective fake names.

The \code{lookup_file} must always be a \code{*.csv}-file with two columns of character vectors, named `real_names` and `fake_names`.

If the specified \code{lookup_file} does not exist, new random `fake_names` are sampled from \pkg{randomNames}, and written to disc as the specified file.
Generated `fake_names` are unique and can be used as R variable names.

If the \code{lookup_file} does not include all \code{real_names}, it is likewise appended with new `fake_names`.
All entries must be unique and valid R variable names.
The rows in the \code{lookup_file} can be in an arbitrary order, and can also include entries that are never used.

By providing a \code{lookup_file} users (or participants) can choose their own `fake_names`, though this may not protect personal data well.
In particular, storing socio-demographic data of participants as custom `fake_names` (such as, for example \code{"m_us_31"}) is not advised, because such data may be easily breached and downstream functions expect socio-demographic data in a different format.
}
\note{
\strong{
  Despite its name, this function \emph{does not magically anonymize data}, but merely replaces names with randomly drawn fake names.
  It is your responsibility to protect your participants' data.
  If you are unsure, or do not understand the below caveats, \emph{do not rely on this function}.
  \enumerate{
    \item The lookup table with real and fake names must be stored in a safe place, ideally \emph{encrypted} and \emph{not together with the raw data or results}.
    \item Your data may still be deanonymized if it includes other personal information and/or few participants.
    \item Make sure no `real names` are included in your command history, caches or other R objects and scripts.
  }
}
}
\examples{
anonymize(real_names = c("Hillary", "Barack", "George"),
          lookup_file = system.file("extdata",
                                    "example_name_lookup.csv",
                                    package = "pensieve"))
# system.file call only necessary for example, shipped with pensieve
# just as an example, never store lookup file with raw data
# see `notes` for details
}
\seealso{
Other import helpers: \code{\link{QItems}},
  \code{\link{check}}
}
\keyword{internal}

