% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{as_psConcourse}
\alias{as_psConcourse}
\title{Coerce objects to psConcourse}
\usage{
as_psConcourse(concourse, type = "text", markup = "plain", babel = TRUE,
  img_dir = NULL, ...)
}
\arguments{
\item{concourse}{A matrix with:
\itemize{
\item researcher-facing \strong{item handles} as row names,
\item \strong{languages} as column names and
\item participant-facing \strong{full items} in cells.
}

For \emph{monolingual studies}, use one column only.

Full items must be unique by language, and can be \code{NA} if not available (not recommended).
Names must be unique and valid R names.}

\item{type}{a character string giving the \emph{kind} of full item stimuli, must be one of:
\itemize{
\item \code{"text"} for textual items, in which case cells in \code{concourse} must be text.
An additional subclass \code{"psConcourseText"} is prepended and validated.
\item \code{"image"} for image items, in which case cells in \code{concourse} must be file paths, relative from \code{img_dir}.
Images must be \code{*.png}, \code{*.jpg}, \code{*.jpeg} or \code{*.svg}.
An additional class \code{"psConcourseImage"} is prepended and validated.
Defaults to \code{"text"}.
}}

\item{markup}{a character string giving the markup for \code{type = "text"}s.
Defaults to \code{"plain"} for plain text.
Currently only allows \code{"plain"}.
Ignored unless \code{type = "text"}.}

\item{babel}{a logical flag, indicating whether \href{https://ctan.org/pkg/babel}{LaTeX's babel package} should be used for multilingual support.
If \code{TRUE} (default), column names in \code{concourse} must be \href{http://ctan.math.washington.edu/tex-archive/macros/latex/required/babel/base/babel.pdf}{valid babel languages}.
Ignored unless \code{type = "text"}.}

\item{img_dir}{a character string giving the directory for \code{type = "image"}s.
Must be relative path \emph{from the working directory}.
Best constructed with \code{\link[base:file.path]{base::file.path()}}.
Defaults to \code{NULL}, in which case images are expected at the working directory root \code{\link[base:getwd]{base::getwd()}}.
Ignored unless \code{type = "image"}.}

\item{...}{additional arguments passed to the S3 method (currently ignored,
except two optional arguments \code{options} and \code{inline}; see
the references below)}
}
\description{
Coerce objects to psConcourse
}
