% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R, R/qItems.R
\name{check}
\alias{check}
\alias{check.qItems}
\title{Validate S3 classes from this package.}
\usage{
check(x)

\method{check}{qItems}(x)
}
\arguments{
\item{x}{[any] class object created by respective \code{make_} function.}
}
\description{
Use \code{check} to validate S3 classed objects from this package.
}
\seealso{
Other import helpers: \code{\link{qItems}}
}

