% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R, R/qItems.R
\name{check}
\alias{assert}
\alias{check}
\alias{check.qItems}
\alias{expect}
\alias{test}
\title{Validate S3 classes from this package.}
\usage{
check(x)

expect(x)

test(x)

assert(x)

\method{check}{qItems}(x)
}
\arguments{
\item{x}{class object created by respective \code{make_} function.}
}
\description{
Use \code{check}, \code{test}, \code{assert} and \code{expect} to validate S3 classed objects from this package.
}
\seealso{
Other import helpers: \code{\link{anonymize}},
  \code{\link{qItems}}
}

