% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QData.R, R/QItems.R, R/check.R, R/correlate.R
\name{check.QPreSorts}
\alias{assert}
\alias{assert.default}
\alias{check}
\alias{check.QCorr}
\alias{check.QItems}
\alias{check.QPeopleFeatures}
\alias{check.QPreSorts}
\alias{check.QSorts}
\alias{expect}
\alias{expect.default}
\alias{test}
\alias{test.default}
\title{Validate S3 classes from this package.}
\usage{
\method{check}{QPreSorts}(x)

\method{check}{QSorts}(x)

\method{check}{QPeopleFeatures}(x)

\method{check}{QItems}(x)

check(x)

test(x)

\method{test}{default}(x)

expect(x, info = NULL, label = NULL)

\method{expect}{default}(x, info = NULL, label = paste(class(x),
  "S3 class"))

assert(x, collection = NULL, var.name = NULL)

\method{assert}{default}(x, collection = NULL, var.name = paste(class(x),
  "S3 class"))

\method{check}{QCorr}(x)
}
\arguments{
\item{x}{class object created by respective constructor function.}

\item{info}{[\code{character(1)}]\cr
See \code{\link[testthat]{expect_that}}}

\item{label}{[\code{character(1)}]\cr
See \code{\link[testthat]{expect_that}}}

\item{collection}{[\code{\link{AssertCollection}}]\cr
If an \code{\link{AssertCollection}} is provided, the error message is stored
in it. If \code{NULL}, an exception is raised if \code{res} is not
\code{TRUE}.}

\item{var.name}{[\code{character(1)}]\cr
The custom name for \code{x} as passed to any \code{assert*} function.
Defaults to a heuristic name lookup.}
}
\description{
Use \code{check}, \code{test}, \code{assert} and \code{expect} to validate S3 classed objects from this package.
}
\seealso{
Other import helpers: \code{\link{QItems}},
  \code{\link{QPeopleFeatures}}, \code{\link{QPreSorts}},
  \code{\link{QSorts}}, \code{\link{anonymize}}
}

