% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{check_latex_lang}
\alias{check_latex_lang}
\title{Check if language can be compiled}
\usage{
check_latex_lang(x)
}
\arguments{
\item{x}{\code{[character(1)} giving a pandoc language, same as \code{lang} in \code{\link[=md2tex]{md2tex()}}.}
}
\value{
\code{TRUE} or error message
}
\description{
Checks if pandoc language can be compiled given local tex distribution
}
\keyword{internal}
