% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_S3}
\alias{check_S3}
\title{Validate S3 classes from pensieve}
\usage{
check_S3(x)
}
\arguments{
\item{x}{class object created by respective constructor function.}
}
\description{
Use \code{check()}, \code{test()}, \code{assert()}, \code{expect()} and \code{need()} to validate  classed objects from this package.

Check S3 class.
}
\details{
\code{check()}, \code{test()}, \code{assert()} and \code{expect()} are extensions the family of functions from the \href{https://github.com/mllg/checkmate}{checkmate} package.
\code{check()} returns \code{TRUE} or the error message, \code{assert()} only returns a message in case of an error, \code{test()} returns \code{TRUE} or \code{FALSE}.
\code{expect()} is for internal use with testing via \href{https://github.com/hadley/testthat}{testthat}).
\code{need()} returns \code{NULL} or the error message and is for interal use with the accio web frontend inside \code{\link[shiny:validate]{shiny::validate()}}.
}
\examples{
# check object and friends ...
# check(x)  # returns TRUE or error message
# test(x)  # returns TRUE or FALSE
# assert(x)  # returns error or silently object
# need(x)  # returns NULL or error message
}
\seealso{
Other validation functions: \code{\link{QSorts}},
  \code{\link{check}}, \code{\link{correlate}},
  \code{\link{extract}}, \code{\link{make_grid}},
  \code{\link{score}}
}
\concept{validation functions}
