% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R, R/psItemContent.R
\name{check_S3}
\alias{check_S3}
\alias{check_S3.psItemContent}
\title{Validate S3 classes from pensieve}
\usage{
check_S3(x, ...)

\method{check_S3}{psItemContent}(x, ...)
}
\arguments{
\item{x}{An object created by soem constructor function.}

\item{...}{further arguments to be passed to methods.}
}
\description{
Use \code{check_S3()}, \code{test()}, \code{assert()}, \code{expect()} and \code{need()} to validate  classed objects from this package.
}
\details{
The validation functions all use the same underlying \code{\link[=check_S3]{check_S3()}} methods, but return their results in one of several forms:
\itemize{
\item \strong{\href{https://github.com/mllg/checkmate}{checkmate}} package extensions:
\itemize{
\item \code{check_S3()} returns \code{TRUE} or the error message as a character string,
\item \code{assert()} returns \code{x} invisibly or throws an error,
\item \code{test()} returns \code{TRUE} or \code{FALSE},
\item \code{expect()} always returns an \code{\link[testthat:expectation]{testthat::expectation()}} for internal use with testing via \href{https://github.com/hadley/testthat}{testthat}).
}
\item \strong{\href{https://shiny.rstudio.com}{shiny}} validation:
\itemize{
\item \code{need()} returns \code{NULL} or the error message for interal use with the accio web frontend inside \code{\link[shiny:validate]{shiny::validate()}}.
}
}
}
\examples{
# just for testing; never build objects like this by hand
good_obj <- structure(.Data = "I am an item.", class = "psItemContent")
bad_obj <- structure(.Data = 1L, class = "psItemContent") # must be character

check_S3(good_obj)
check_S3(bad_obj)
}
\concept{validation functions}
