% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{civicon_2014}
\alias{civicon_2014}
\title{Q-Sorts on taxation and the economy from the 2014 CiviCon Citizen Conference}
\format{A list with 3 elements:
test}
\source{
\url{http://www.maxheld.de/schumpermas/}
}
\usage{
civicon_2014
}
\description{
A dataset of 18 participants sorting 77 items on taxation and the economy before and after the 2014 CiviCon Citizen Conference.
}
\keyword{datasets}

