% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlate.R
\name{correlate}
\alias{correlate}
\alias{QCorr}
\title{Correlate Q Sorts.}
\usage{
correlate(sorts, algorithm = "spearman", use = "pairwise.complete.obs")
}
\arguments{
\item{sorts}{an integer matrix with item-cases as named rows, people-variables as named columns and item positions in cells.
Higher-dimensionality objects fom \code{\link{QSorts}} must be subset.}

\item{algorithm}{choice from \code{'pearson'}, \code{'spearman'} or \code{'kendall'}, passed on to \code{\link[stats]{cor}}.
Defaults to \code{'spearman'}.}

\item{use}{an optional character string giving a
    method for computing covariances in the presence
    of missing values.  This must be (an abbreviation of) one of the strings
    \code{"everything"}, \code{"all.obs"}, \code{"complete.obs"},
    \code{"na.or.complete"}, or \code{"pairwise.complete.obs"}.}
}
\value{
A numerical matrix of class \code{QCorr}.
}
\description{
Correlate Q Sorts.
}
\examples{
correlate(sorts = civicon_2014$qData$sorts[,,"before"])

}
\seealso{
Other correlation functions: \code{\link{plot.QCorr}}
}
