% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_data.R
\name{diag_distro_max}
\alias{diag_distro_max}
\title{Infer maximum grid of q-sorting grid.}
\usage{
diag_distro_max(sorts)
}
\arguments{
\item{sorts}{An integer array with item handles as first dimension, people as second dimension, arbitrary dimensions thereafter, and item positions in cells.
Dimensions must be named.}
}
\value{
A positive integer vector of counts of length equal to cover the range of values, named by values.
}
\description{
Infers the maximum q-sorting grid from observed dataset.
}
\examples{
diag_distro_max(sorts = civicon_2014$qData$sorts[,,"before"])
# true, because study used forced Q distribution

}
\seealso{
Other distribution helpers: \code{\link{diag_distros}},
  \code{\link{diag_inside_grid}}, \code{\link{diag_same}}
}
\keyword{internal}
