% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_data.R
\name{diag_forced}
\alias{diag_forced}
\title{Forced distribution test.}
\usage{
diag_forced(dataset, grid)
}
\arguments{
\item{dataset}{An integer matrix with item-cases as rows and people-variables as columns.}

\item{grid}{A positive integer vector of a length covering the range of values, specifying maximum allowed counts for each value.
(in Q-parlance, the maximum column heights for the Q-sorts).}
}
\value{
A logical value of length 1, \code{TRUE} if the distribution is forced, else \code{FALSE}.
}
\description{
Test whether all people-variables completely fill out the same grid.
}
\details{
If all people-variables display the same value counts, respondents plausibly faced a \emph{forced} distribution, but this does not strictly speaking follow.
It is unlikely, though conceivable, that respondents all adhered to precisely the same distribution, but \emph{could} have used other available slots.

Whether, or not, a Q dataset should be considered "forced" can therefore only ascertained from a comparison with the physically available grid during the administration of the Q-sort.

In the above, improbable case, the function warns the user of possible misspecification.
}
\examples{
dataset <- civicon_2014$sorts[,,"before"]
diag_forced(dataset = dataset,
            grid = diag_distro_max(dataset))

}
\seealso{
Other distribution diagnostic functions: \code{\link{diag_distro_max}},
  \code{\link{diag_distros}},
  \code{\link{diag_inside_grid}}, \code{\link{diag_same}}
}
\keyword{internal}

