% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diag_inside_grid}
\alias{diag_inside_grid}
\title{Grid test.}
\usage{
diag_inside_grid(sorts, grid)
}
\arguments{
\item{sorts}{An integer array with item handles as first dimension, people as second dimension, arbitrary dimensions thereafter, and item positions in cells.
Dimensions must be named.}

\item{grid}{A positive integer vector of a length covering the range of values, specifying maximum allowed counts for each value.
(in Q-parlance, the maximum column heights for the Q-sorts).}
}
\value{
A logical vector of length 1, \code{TRUE} if all people-variables fall inside the grid, else \code{FALSE}.
}
\description{
Test whether the value counts of all people-variables fall inside the grid.
}
\examples{
sorts <- civicon_2014$qData$sorts[,,"before"]
diag_inside_grid(sorts = sorts,
                 grid = diag_distro_max(sorts))
# circularly true!

}
\seealso{
Other distribution helpers: \code{\link{diag_distro_max}},
  \code{\link{diag_distros}}, \code{\link{diag_same}}
}
\concept{distribution helpers}
\keyword{internal}
