% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{extract}
\alias{extract}
\alias{QLoas}
\alias{check.QLoas}
\alias{plot.QLoas}
\title{Dimensionality reduction.}
\usage{
extract(cors, nfactors, fa_type = "pca")

QLoas(loas, validate = TRUE)

\method{check}{QLoas}(x)

\method{plot}{QLoas}(x, summarize = NULL, by = "people", r2 = FALSE,
  use_js = NULL, ...)
}
\arguments{
\item{cors}{A numerical matrix with correlations.}

\item{nfactors}{Positive, non-zero integer scalar, giving the number of components (or factors) to retain.
Must be smaller than the number of observations (here: items) on which \code{cors} are based.}

\item{fa_type}{String giving the type of factor analysis performed.
Currently only \code{"pca"} for principal components analysis (PCA) is allowed.}

\item{loas}{A numerical matrix with loadings.}

\item{validate}{A logical flag, indicating whether the object will be validated on construction.
Defaults to `TRUE`.}

\item{x}{class object created by respective constructor function.}

\item{summarize}{A logical flag, indicating whether the object should be summarized before plotting.
If \code{TRUE}, a plot with a density estimate of individual data points is returned, suitable for medium to large objects (numbers of people-variables).
If \code{FALSE}, a plot with individual data points is returned, suitable for small to medium-sized objects (numbers of people-variables).

Defaults to \code{NULL}, in which case the appropriate plot is inferred from the size.}

\item{by}{A character string indicating by which dimension to plot.
Must be \code{"people"} or \code{"both"}.
Display by \code{"factor"} is not currently available.}

\item{r2}{Logical flag, indicating whether the \emph{squared} loadings should be plotted.
If \code{FALSE}, loadings are plotted, interpretable as correlations of people-variables with factors.
If \code{TRUE}, squared loadings are plotted, interpretable as the share of explained variance.
Defaults to \code{FALSE}.}

\item{use_js}{A logical flag, indicating whether an interactive, java-script variant of the plot should be returned.
Defaults to `NULL`, in which case the appropriate output is inferred from the runtime environment.}

\item{...}{Arguments passed onto other methods.
Not currently used.}
}
\description{
Wraps Principal Components Analysis (PCA) to extract component loadings.

Construct S3 class.

Check S3 class.

Plot S3 class.
}
\section{Methods (by generic)}{
\itemize{
\item \code{check}: validation

\item \code{plot}: plotting
}}

\examples{
# this just runs the calculations, but also classes and validates the results
cors <- correlate(sorts = civicon_2014$qData$sorts[,,"before"])  # preliminary step
loas <- extract(cors = cors, nfactors = 3, fa_type = "pca")
# this just assigns the class, without validation (not recommended)
loas <- QLoas(loas = loas, validate = FALSE)
# this validates the class
check(loas)
# makes a barchart
plot(x = loas, use_js = NULL, by = "people")
plot(x = loas, use_js = NULL, by = "both")

}
\seealso{
Other analysis functions: \code{\link{correlate}},
  \code{\link{score}}

Other S3 classes from `pensieve`: \code{\link{QConcourse}},
  \code{\link{QItemStrata}}, \code{\link{QPeopleFeatures}},
  \code{\link{QPreSorts}}, \code{\link{QSorts}},
  \code{\link{correlate}}, \code{\link{make_grid}},
  \code{\link{score}}

Other validation functions: \code{\link{QSorts}},
  \code{\link{check}}, \code{\link{correlate}},
  \code{\link{make_grid}}, \code{\link{score}}

Other plotting functions: \code{\link{QSorts}},
  \code{\link{correlate}},
  \code{\link{knit_print.QConcourse}}
}
