% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markup2vector.R
\name{find_fontsize}
\alias{find_fontsize}
\title{Find largest possible fontsize given all other arguments}
\usage{
find_fontsize(l, fontsizes_local_possible = fontsizes_local, ...)
}
\arguments{
\item{l}{\code{[list()]}
giving \code{x}s to be passed to \code{\link[=md2tex_mem]{md2tex_mem()}}.}

\item{fontsizes_local_possible}{\code{[character()]} giving possible fontsizes_local, defaults to all allowed values as per \code{\link[=md2tex_mem]{md2tex_mem()}}.}

\item{...}{arguments passed on to downstream formatting functions}
}
\value{
\code{[character(1)]} giving largest possible fontsize
}
\description{
Finds largest possible fontsize for list of markdown vectors to fit on one PDF page.
}
\keyword{internal}
