% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{find_max_fontsize}
\alias{find_max_fontsize}
\title{Find largest possible fontsize given all other arguments}
\usage{
find_max_fontsize(fontsizes_local_possible = fontsizes_local, ...)
}
\arguments{
\item{fontsizes_local_possible}{\code{[character()]} giving possible fontsizes_local}

\item{...}{Arguments passed on to \code{md2tex_mem}
\describe{
  \item{x}{\code{[character()]} \emph{or} \code{[raw()]} giving the input.}
  \item{path_in}{\code{[character(1)]} giving path to use for input file \emph{with or without extension}.
Defaults to \code{"foo"}.
Useful for debugging.}
  \item{alignment}{\code{[character(1)]} giving the alignment of the text.

Must be one of:
\itemize{
\item \code{'justified'} (\strong{default})
\item \code{'left'}
\item \code{'right'} or
\item \code{'center'}
}}
}}
}
\value{
\code{[character(1)]} giving largest possible fontsize
}
\description{
Finds largest possible fontsize for some markdown to fit on one PDF page.
}
\keyword{internal}
