% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inanimatus.R
\name{inanimatus}
\alias{inanimatus}
\title{HTML widget factory for sorting grids}
\usage{
inanimatus(grid = as_psGrid(obj = c(1, 2, 3, 5, 3, 2, 1)),
  sort = matrix(data = c(NA, NA, NA, NA, "foo", NA, NA, NA, "zap", "bar",
  NA, NA, "hup", "op", "schwupp", "rip", "zonk", "oupsi", "what", "omg",
  NA, NA, "zor", "arr", "apple", NA, NA, NA, "warp", "glips", NA, NA, NA,
  NA, "ronk"), ncol = 7), header = TRUE, footer = TRUE,
  aspect_ratio_cards = 16/9, width = NULL, height = NULL)
}
\arguments{
\item{grid}{\code{[matrix()]}
giving the availability of cells for sorting as \code{logical(1)} values.
At least one dimension should be named (see examples), or the x-axis (column names) is assumed to be the sorting direction.
Unnamed dimensions are assumed to be meaningless, i.e. used for stacking tied items.}

\item{sort}{\code{[matrix()]}
giving the occupying item of cells as \code{character(1)} strings of \strong{item handles}.
\code{NA} is used for empty \emph{and} disallowed cells (see \link{psGrid}).
At least one dimension should be named (see examples), or the x-axis (column names) is assumed to be the sorting direction.
Unnamed dimensions are assumed to be meaningless, i.e. used for stacking tied items.}

\item{header}{A logical flag, defaults to \code{TRUE}, in which case column names  from \code{grid} are included as headers.}

\item{footer}{A logical flag, defaults to \code{TRUE}, in which case column names  from \code{grid} are included as footers.}

\item{aspect_ratio_cards}{A numeric scalar, giving width divided by height of \emph{individual cards} (such as 16/9 for screen dimensions).
Aspect ratio of \emph{cards} is required to appropriately set the resulting dimensions of the \emph{grid}.
Defaults to standard business cards.}

\item{width}{Fixed width for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{height}{Fixed height for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}
}
\value{
An htmlwidget.
}
\description{
Creates HTML widgets from sorting grids.
Wraps pensieve's inanimatus.js.
End users should use the convenient print method wrappers.
}
