% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inanimatus.R
\name{inanimatus}
\alias{inanimatus}
\title{HTML widget factory for sorting grids}
\usage{
inanimatus(grid = as_psGrid(obj = c(1, 2, 3, 5, 3, 2, 1)), header = TRUE,
  footer = TRUE, aspect_ratio_cards = 16/9, width = NULL, height = NULL)
}
\arguments{
\item{grid}{\code{[matrix()]}
giving the availability of cells for sorting as \code{logical(1)} values.
The (horizontal) x-axis is assumed to be the sorting direction, the (vertical) y-axis for recording ties.
Dimensions can be named (recommended), giving a short description of the sorting dimension (only applicable to the x-axis).
Row and column \emph{indeces} can also be named, but names are purely cosmetic.}

\item{header}{A logical flag, defaults to \code{TRUE}, in which case column names  from \code{grid} are included as headers.}

\item{footer}{A logical flag, defaults to \code{TRUE}, in which case column or row names from \code{grid} are included as headers or footers, respectively.
Missing row or column names are replaced with sensible defaults.}

\item{aspect_ratio_cards}{A numeric scalar, giving width divided by height of \emph{individual cards} (such as 16/9 for screen dimensions).
Aspect ratio of \emph{cards} is required to appropriately set the resulting dimensions of the \emph{grid}.
Defaults to standard business cards.}

\item{width}{Fixed width for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{height}{Fixed height for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}
}
\value{
An htmlwidget.
}
\description{
Creates HTML widgets from sorting grids.
Wraps pensieve's inanimatus.js.
End users should use the convenient print method wrappers.
}
