% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{itemConcourse}
\alias{itemConcourse}
\title{Construct itemConcourse}
\usage{
itemConcourse(concourse = NULL, languages = c("english"),
  type = "textItem")
}
\arguments{
\item{concourse}{\itemize{
\item for \emph{monolingual studies}, when only one \code{languages} is given: a named character vector of (participant-facing) \strong{full items}, with names as (researcher-facing) \strong{item handles},
\item for \emph{multilingual studies}, when several \code{languages} are given: a character matrix of (participant-facing) \strong{full items}, with rownames as (researcher-facing) \strong{item handles}.
}

Full items must be unique by language, and can be \code{NA} if not available (not recommended).
Names must be unique and valid R names.}

\item{languages}{a character vector of languages, giving the column names.
Defaults to \code{c("english")} for a single language concourse in english.
Must be unique and valid R names.}

\item{type}{a character string giving the \emph{kind} of item stimuli, must be one of:
\itemize{
\item \code{"textItem"} for textual items, in which case cells in \code{concourse} must be plain text.
An additional subclass \code{"textItem"} is prepended and validated.
\item \code{"imageItem"} for image items, in which case cells in \code{concourse} must be file paths, relative from the current working directory.
An additional class \code{"imageItem"} is prepended and validated.
Defaults to \code{"text"}.
}}
}
\description{
Stores \emph{all} (researcher-facing) \strong{item handles} and (participant-facing) \strong{full items}.
}
\examples{
# simpler syntax for monolingual studies
x <- itemConcourse(
  concourse = c(live_2_work = "Man lives to work.",
                work_2_live = "Man works to live."),
  languages = c("english"),
  type = "textItem"
)


# more complex interface for multilingual studies
x <- itemConcourse(
  concourse = matrix(
    data = c(
      "Man lives to work.", "Man lebt, um zu arbeiten.",
      "Man works to live.", "Man arbeitet, um zu leben."
    ),
    nrow = 2,
    ncol = 2,
    dimnames = list(items = c("live_2_work", "work_2_live"))),
  languages = c("english", "ngerman"),
  type = "textItem"
)

}
