% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QItems.R
\name{knit_print.QConcourse}
\alias{knit_print.QConcourse}
\title{Custom print method for knitr}
\usage{
knit_print.QConcourse(x, use_js = NULL, ...)
}
\arguments{
\item{x}{a character matrix with full item wording of class \code{\link{QConcourse}}, as created by \code{\link[=QConcourse]{QConcourse()}}.}

\item{use_js}{A logical flag, indicating whether an interactive, java-script variant of the plot should be returned.
Defaults to `NULL`, in which case the appropriate output is inferred from the runtime environment.}

\item{...}{additional arguments passed to the S3 method (currently ignored,
except two optional arguments \code{options} and \code{inline}; see
the references below)}
}
\description{
Provides custom print method for knitr.
Can also be invoked manually to open interactive outputs in RStudio.
}
\seealso{
Other printing functions: \code{\link{plot.QCors}}
}
