% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{komki}
\alias{komki}
\title{Q-Sorts and Q-Categorisations on language games}
\format{A list with 3 elements:
\describe{
  \item{items}{a dataframe with german and english handles,}
  \item{qsorts}{a matrix with items as rows, participants as columns and Q-sorts in cells.}
  \item{qcat}{Q categorisation in its canonical form.}
}}
\source{
\url{https://www.erziehungswissenschaften.hu-berlin.de/de/institut/mitarbeiter/1688111}
}
\usage{
komki
}
\description{
A dataset of 53 participants, sorting and categorizing 35-41 items.
Part of Verena Kasztantowicz's dissertation at Humboldt University of Berlin, Germany.
}
\keyword{datasets}
