% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design.R
\name{make_grid}
\alias{make_grid}
\alias{QGrid}
\alias{check.QGrid}
\title{Create Q sorting grid.}
\usage{
make_grid(x_range = c(-5, 5), y_range = c(1, 5), pattern = "honeycomb",
  offset = "odd")

QGrid(grid, pattern = "honeycomb", offset = "even", validate = TRUE)

\method{check}{QGrid}(x)
}
\arguments{
\item{x_range}{An integer vector of length two, giving the range on the x axis.
Defaults to \code{c(-5, 5)}.}

\item{y_range}{An integer vector of length two, giving the range on the y axis.
Defaults to \code{c(1, 5)}.}

\item{pattern}{A character string, giving the pattern of tesselation to use.
Must be \code{"honeycomb"} (default), the only currently supported pattern.
Defaults to \code{NULL}, in which case \code{pattern} is expected as an attribute of \code{grid}.}

\item{offset}{A character string, giving the rows to be offset.
Must be \code{"even"} or \code{"odd"} (default).
Applies only to \code{"honeycomb"} and \code{"brickwall"} patterns.}

\item{grid}{A logical matrix giving the available cells for Q sorting.}

\item{validate}{a logical flag, indicating whether the object will be validated on construction.
Defaults to \code{TRUE}.}

\item{x}{class object created by respective constructor function.}
}
\description{
Creates a grid for Q sorting.

Construct S3 class for pensieve.

Check S3 class.
}
\details{
\code{check()}, \code{test()}, \code{assert()} and \code{expect()} are extensions the family of functions from the \href{https://github.com/mllg/checkmate}{checkmate} package.
\code{check()} returns \code{TRUE} or the error message, \code{assert()} only returns a message in case of an error, \code{test()} returns \code{TRUE} or \code{FALSE}.
\code{expect()} is for internal use with testing via \href{https://github.com/hadley/testthat}{testthat}).
\code{need()} returns \code{NULL} or the error message and is for interal use with the accio web frontend inside \code{\link[shiny:validate]{shiny::validate()}}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{check}: validation
}}

\examples{
# this makes the grid, already assigns class
grid <- make_grid(x_range = c(-5, 5), y_range = c(1,5), pattern = "honeycomb", offset = "odd")
# make simple matrix by hand
m <- matrix(data = c(TRUE, FALSE, FALSE, TRUE), nrow = 2, ncol = 2)
# assign class, without validation (not recommended)
grid <- QGrid(grid = m, pattern = "honeycomb", offset = "even", validate = FALSE)
# check object and friends ...
# check(x)  # returns TRUE or error message
# test(x)  # returns TRUE or FALSE
# assert(x)  # returns error or silently object
# need(x)  # returns NULL or error message
# validate the class
check(grid)
}
\seealso{
Other S3 classes from `pensieve`: \code{\link{QPreSorts}},
  \code{\link{QSorts}}, \code{\link{correlate}},
  \code{\link{extract}}, \code{\link{psConcourse}},
  \code{\link{psItems}}, \code{\link{psOpenSorts}},
  \code{\link{psOpenSort}}, \code{\link{psPeople}},
  \code{\link{score}}

Other validation functions: \code{\link{QSorts}},
  \code{\link{check}}, \code{\link{correlate}},
  \code{\link{extract}}, \code{\link{score}}
}
\concept{S3 classes from `pensieve`}
\concept{design functions}
\concept{survey functions}
\concept{validation functions}
