% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{pensieveItems}
\alias{pensieveItems}
\title{Construct list of item-related objects}
\usage{
pensieveItems(pensieve_concourse)
}
\arguments{
\item{pensieve_concourse}{object returned by \code{\link[=pensieveConcourse]{pensieveConcourse()}}}
}
\description{
Constructs S3 class for pensieve.
}
\seealso{
Other S3 classes from `pensieve`: \code{\link{QItemStrata}},
  \code{\link{QPreSorts}}, \code{\link{QSorts}},
  \code{\link{correlate}}, \code{\link{extract}},
  \code{\link{make_grid}}, \code{\link{pensieveConcourse}},
  \code{\link{pensieveOpenSorts}},
  \code{\link{pensievePeople}}, \code{\link{score}}
}
