% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psClosedSorts.R
\name{psClosedSorts}
\alias{psClosedSorts}
\alias{validate_S3.psClosedSorts}
\alias{as_psClosedSorts}
\alias{as_psClosedSorts.psSort}
\alias{plot.psClosedSorts}
\title{Store multiple sorts as a numeric matrix}
\usage{
psClosedSorts(csorts)

\method{validate_S3}{psClosedSorts}(x, items = NULL, grid = NULL, ...)

as_psClosedSorts(obj, ...)

\method{as_psClosedSorts}{psSort}(obj, ...)

\method{plot}{psClosedSorts}(x, column = 1, use_js = NULL, ...)
}
\arguments{
\item{csorts}{\code{[array()]}
An numeric array with people as rows, item handles as columns, arbitrary dimensions thereafter (such as for multiple sorting dimensions), and item positions in cells.}

\item{x}{An object with one of the pensieve S3 classes.}

\item{items}{\code{[character()]} giving the \emph{participant-facing} \strong{item content}.
Can be named to provide short, \emph{researcher-facing} \strong{item handles}.
\itemize{
\item if \code{dir_bin} is \code{NULL} (default), \code{items} must be text.
Items can be marked up using \href{https://rmarkdown.rstudio.com/authoring_pandoc_markdown.html}{Pandoc Markdown}.
An additional subclass \code{psItemContentText} is prepended and validated.
\item if \code{dir_bin} is given, \code{items} must be file paths, relative from \code{dir_bin}.
An additional subclass \code{psItemContentBin`` is prepended and validated.}lang\code{,}fontsize_global\code{,}alignment\code{and}linestretch` are ignored.
}}

\item{grid}{\code{[matrix()]}
giving the availability of cells for sorting as \code{logical(1)} values.
The (horizontal) x-axis is assumed to be the sorting direction, the (vertical) y-axis for recording ties.
Dimensions can be named (recommended), giving a short description of the sorting dimension (only applicable to the x-axis).
Row and column \emph{indeces} can also be named, but names are purely cosmetic.}

\item{...}{Arguments passed onto other methods.
Not currently used.}

\item{obj}{An object which can be coerced to an integer array of class \link{psClosedSorts}.}

\item{column}{Positive integer scalar, giving the column of the psClosedSorts object to plot.
Defaults to \code{1}, in which case the first column is plotted.}

\item{use_js}{A logical flag, indicating whether an interactive, java-script variant of the plot should be returned.
Defaults to \code{NULL}, in which case the appropriate output is inferred from the runtime environment.}
}
\value{
A numeric matrix of class \link{psClosedSorts}.
}
\description{
Canonical storage for closed sorts.

Plot S3 class.
}
\section{Methods (by generic)}{
\itemize{
\item \code{validate_S3}: Validation against items and grid (recommended)

\item \code{as_psClosedSorts}: Coercion from \link{psSort} (creates one row)

\item \code{plot}: plotting
}}

\examples{
# multiple sorts ====
# simple case with one condition of instruction
csorts <- matrix(
  data = c(-1, 0, 1, -1),
  nrow = 2,
  byrow = TRUE,
  dimnames = list(
    people = c("Lisa", "Peter"),
    items = c("live_2_work", "work_2_live")
  )
)
csorts <- psClosedSorts(csorts)

# two conditions of instruction are planes in a 3-d array
csorts_multiple_conds <- array(
  data = c(-1, 0, 1, -1, 0, 1, -1, 0),
  dim = c(2, 2, 2),
  dimnames = list(
    people = c("Lisa", "Peter"),
    items = c("live_2_work", "work_2_live"),
    conditions = c("desirable", "probable")
  )
)
csorts_multiple_conds <- psClosedSorts(csorts_multiple_conds)
}
\seealso{
Other S3 classes from `pensieve`: \code{\link{correlate}},
  \code{\link{extract}}, \code{\link{psGrid}},
  \code{\link{psItemContent}}, \code{\link{psOpenSorts}},
  \code{\link{psOpenSort}}, \code{\link{psPeople}},
  \code{\link{score}}

Other plotting functions: \code{\link{correlate}},
  \code{\link{extract}}
}
\concept{S3 classes from `pensieve`}
\concept{plotting functions}
