% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{psItems}
\alias{psItems}
\title{Construct list of item-related objects}
\usage{
psItems(ps_concourse)
}
\arguments{
\item{ps_concourse}{object returned by \code{\link[=psConcourse]{psConcourse()}}}
}
\description{
Constructs S3 class for pensieve.
}
\seealso{
Other S3 classes from `pensieve`: \code{\link{QItemStrata}},
  \code{\link{QPreSorts}}, \code{\link{QSorts}},
  \code{\link{correlate}}, \code{\link{extract}},
  \code{\link{make_grid}}, \code{\link{pensievePeople}},
  \code{\link{psConcourse}}, \code{\link{psOpenSorts}},
  \code{\link{score}}
}
