% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psOpenSort.R
\name{psOpenSort}
\alias{psOpenSort}
\alias{psLogicalOpenSort}
\alias{psOrdinalOpenSort}
\alias{psIntervalOpenSort}
\alias{as_psOpenSort}
\alias{as_psOpenSort.matrix}
\alias{as_psOpenSort.data.frame}
\alias{tidy.psLogicalOpenSort}
\alias{autoplot.psLogicalOpenSort}
\alias{summary.psLogicalOpenSort}
\title{Store \emph{individual} sort and dimension description by \emph{one} participant.}
\usage{
psOpenSort(osort, descriptions = NULL, scale = NULL)

as_psOpenSort(osort, descriptions = NULL, scale = NULL)

\method{as_psOpenSort}{matrix}(osort, descriptions = NULL,
  scale = NULL)

\method{as_psOpenSort}{data.frame}(osort, descriptions = NULL,
  scale = NULL)

\method{tidy}{psLogicalOpenSort}(x)

\method{autoplot}{psLogicalOpenSort}(object, edge_codings = NULL,
  str_wrap_width = 30)

\method{summary}{psLogicalOpenSort}(object, ...)
}
\arguments{
\item{osort}{a matrix with items as rows, open dimensions as columns, and open sort value in cells.

If rows are named (by item handles), names must be valid R names.

If columns are named, they must be valid R names and they must be the same as the names in \code{descriptions}.
Either way, \code{osort} and \code{descriptions} are always \emph{matched by index only}: the first column from \code{osort}, must be the first element of \code{description}, and so forth.}

\item{descriptions}{a character vector giving the open-ended dimension description provided by the participant.
If elements are named, names must be valid R names.
Defaults to \code{NULL}, in which case no user-provided dimension descriptions are available (not recommended).}

\item{scale}{a charater string giving the \emph{scale} of the open sorts, must be one of:
\itemize{
\item \code{logical} for \emph{nominally}-scaled sort, where an open dimension applies (\code{TRUE}) or does not apply (\code{FALSE}).
\code{osort} must be a logical matrix.
The subclass \code{"psLogicalOpenSort"} is prepended and validated.
\item \code{ordinal} for an \emph{ordinally}-scaled sort, where an open dimension applies to some item \emph{more} (\code{2nd} rank) \emph{or less} (\code{3rd} rank) than to another other item.
\code{osort} must be an \code{integer} matrix.
The subclass \code{"psOrdinalOpenSort"} is prepended and validated.
\item \code{interval} for \emph{interval} or \emph{ratio}-scaled sort, where an open dimension applies to some item \emph{by some amount more or less} (say \code{2.4} units) than to another item.
\code{osort} must be a \code{numeric} matrix.
The subclass \code{"psIntervalOpenSort"} is prepended and validated.
}

Defaults to \code{NULL}, in which case the scale is inferred from the implicit class of \code{osort}.

Currently only \code{logical} is supported.}

\item{x}{a \link{psLogicalOpenSort}, created by \code{\link[=psOpenSort]{psOpenSort()}}.}

\item{object}{a \link{psLogicalOpenSort}, created by \code{\link[=psOpenSort]{psOpenSort()}}.}

\item{edge_codings}{a tibble with category description indeces in the first column, and arbitrary metadata \emph{about the descriptions} in limited later columns:
\itemize{
\item 2nd column will be mapped to line color,
\item 3rd column will be mapped to line type (and must be discrete).
Later columns will be ignored.
Useful if participants or researchers have coded the open-ended descriptions in some way.
Category description indeces must be a subset of the column names in \code{x}.
If more then one code applies to a category, multiple rows with identical categories can exist, and multiple, "fanned-out" edges will be drawn between the respective nodes.
See note.
Defaults to \code{NULL}.
}}

\item{str_wrap_width}{integer scalar, giving the maximum number of characters after which to insert a newline, passed on to \code{\link[stringr:str_wrap]{stringr::str_wrap()}}.
Defaults to \code{30}.
Useful for long descriptions.}

\item{...}{further arguments passed to methods.}
}
\value{
Object of class \code{psOpenSort}.
}
\description{
Constructs S3 class for pensieve.
}
\section{Methods (by generic)}{
\itemize{
\item \code{as_psOpenSort}: coerce matrix to psOpenSort

\item \code{as_psOpenSort}: coerce data.frame to psOpenSort

\item \code{tidy}: Prepare \emph{individual} open sort for bipartite plotting.

\item \code{autoplot}: Create bipartite plot from \emph{individual} open sort.

\item \code{summary}: Summarise an \emph{individual} open sort.
}}

\note{
To render the resulting \code{ggplot} object, you must \emph{manually} call \code{library(ggraph)} somewhere in your script (as per \href{https://github.com/thomasp85/ggraph/issues/85}{this limitation}).

If \code{codings} are added, the aesthetics are set \emph{for each individual \link{psOpenSort} separately}, which may make it hard to compare plots across participants.
To get consistent code aesthetics, consider applying \code{\link[ggraph:scale_edge_color_manual]{ggraph::scale_edge_color_manual()}} and friends.
}
\examples{
# create single open sort ====

# Lisas open sort, unnamed descriptions (matched by index)
losort <- matrix(
  data = c(TRUE, FALSE, FALSE, FALSE, TRUE, FALSE),
  nrow = 3,
  dimnames = list(items = c("cat", "dog", "cow")))
descriptions <- c(
  "a pet which largely takes care of itself",
  NA  # dimension is assigned, but not described (not a problem)
)
lisa <- psOpenSort(osort = losort, descriptions = descriptions)

# Peters open sort, named descriptions (*also* only matched by index)
losort <- matrix(
  data = c(TRUE, TRUE, FALSE, TRUE, FALSE, FALSE, FALSE, FALSE, TRUE),
  nrow = 3,
  dimnames = list(
    items = c("cat", "dog", "cow"),
    categories = c("in_homes", "quiet", "herbivore")
  ))
descriptions <- c(
  in_homes = "Animal found in peoples homes.",
  quiet = "Does not make a lot of noise.",
  herbivore = "Eats plants.")  # defined, but never TRUE (not a problem)
peter <- psOpenSort(osort = losort, descriptions = descriptions)

# coercion methods
peter_m <- as_psOpenSort(osort = as.matrix(x = losort), descriptions = descriptions)
peter_df <- as_psOpenSort(osort = as.data.frame(x = losort), descriptions = descriptions)

# Rebeccas open sort, without any descriptions provided
losort <- matrix(
  data = c(FALSE, FALSE, TRUE, TRUE, TRUE, FALSE),
  nrow = 3,
  dimnames = list(handles = c("cat", "dog", "cow")))
rebecca <- psOpenSort(osort = losort, descriptions = NULL)

# Ira open sort, with some problems
losort <- matrix(
  data = c(
    FALSE, FALSE, FALSE,  # this is dropped, b/c there is just no valuable information here,
    TRUE, TRUE, TRUE,  # same problem; no variance
    FALSE, FALSE, FALSE,
    # also no variance, but there *is* a corresponding description,
    # so we're setting column to NA and keeping the description
    NA, TRUE, FALSE),  # you can also have *actual* NAs
  nrow = 3,
  byrow = FALSE,
  dimnames = list(handles = c("cat", "dog", "cow"))
)
descriptions <- c(NA, NA, "mammals", NA)
ira <- suppressWarnings(as_psOpenSort(osort = losort, descriptions = descriptions))
# this gives appropriate warning messages
# psOpenSort() would error out; only coercion method will attempt fix

# ordinally and intervally scaled sorts are also possible, but currently unsupported
tyler <- matrix(
  data = as.integer(c(1, 2, 2, 1)),
  nrow = 2,
)
tyler <- psOpenSort(
  osort = tyler,
  scale = "ordinal")  # defaults to implicit class of base type
roberta <- matrix(
  data = c(2.2, 4.3, -2.8, 0),
  nrow = 2
)
roberta <- psOpenSort(osort = roberta)

# plotting ====
ggplot2::autoplot(object = lisa)
ggplot2::autoplot(object = rebecca)

# no with codes
petercodes <- tibble::tibble(category = c("in_homes", "in_homes", "quiet", "herbivore"),
                             reference = c("location", "human interaction", NA, "animal diet"),
                             length = c("medium", "medium", "medium", "short")
                             # notice the duplicates to allow for multiple codes
                             )
library(ggraph)  # must be attached while running below
ggplot2::autoplot(object = peter, edge_codings = petercodes)
summary(peter)

}
\seealso{
Other S3 classes from `pensieve`: \code{\link{correlate}},
  \code{\link{extract}}, \code{\link{psClosedSorts}},
  \code{\link{psGrid}}, \code{\link{psItemContent}},
  \code{\link{psOpenSorts}}, \code{\link{psPeople}},
  \code{\link{score}}
}
\concept{S3 classes from `pensieve`}
