% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psOpenSorts.R
\name{psOpenSorts}
\alias{psOpenSorts}
\alias{import_psOpenSorts}
\alias{psOpenSort}
\alias{tidy.psOpenSort}
\alias{autoplot.psOpenSort}
\alias{summary.psOpenSort}
\alias{tidy.psOpenSorts}
\alias{autoplot.psOpenSorts}
\title{Store open sorts in a psOpenSort list.}
\usage{
psOpenSorts(open_sorts)

import_psOpenSorts(assignments_messy, descriptions_messy = NULL)

psOpenSort(assignments, descriptions = NULL)

\method{tidy}{psOpenSort}(x)

\method{autoplot}{psOpenSort}(object, edge_codings = NULL,
  str_wrap_width = 30)

\method{summary}{psOpenSort}(object, ...)

\method{tidy}{psOpenSorts}(x)

\method{autoplot}{psOpenSorts}(object)
}
\arguments{
\item{open_sorts}{A list of matrices, one for each participant.
Matrices must be \code{\link[=psOpenSort]{psOpenSort()}} object, or coercable to \code{\link[=psOpenSort]{psOpenSort()}}.}

\item{assignments_messy}{a character matrix with rows as items, columns as participants and  \strong{logical category assignments} as character strings in cells.
Categories are identified by a subset from \code{LETTERS}, same as in \code{descriptions_messy}.
Assignments must be the same subset of \code{LETTERS} as the column names in \code{descriptions_messy}.
Rows and columns must be named.

For example, if some participant assigned her (self-described) categories \code{A}, \code{D} and \code{Z} to some item, the cell for that item and participant would read \code{"A, D, Z"}.
Order and punctuation are ignored.

See \code{note}.}

\item{descriptions_messy}{a character matrix with rows as category indices, columns as participants and \strong{category descriptions} in cells.
Rows \emph{must} be named by a subset of \code{LETTERS} to conveniently enter, and identify them from \code{assignments_messy}.
The row names are arbitrary identifiers, but will be retained for the canonical form.
Columns \emph{must} be named as participants.

Defaults to \code{NULL}, in which case no descriptions are available.

Notice category description in one row have \emph{nothing in common} other than their \emph{indices}:
For example, the category descriptions in a row named \code{'B'} are all by different participants, and may refer to entirely different aspects.
They are only conveniently entered in a table, and all share the fact that they were the \emph{second} description provided.

When some category has not been defined by the participant, the value in the cell should be \code{NA}.
Empty strings \code{""} will also be considered \code{NA}.}

\item{assignments}{a matrix with item-handles as row names, arbitrary or empty column names, and open sort value in cells.
Matrix must be either
\itemize{
\item \code{logical} for \emph{nominal}-scaled sort, where an open category applies (\code{TRUE}) or does not apply (\code{FALSE}),
\item \code{integer} for \emph{ordinally}-scaled sort, where an open category applies to some item \emph{more} (\code{2nd} rank) \emph{or less} (\code{3rd} rank) than to another other item,
\item \code{numeric} for \emph{interval} or \emph{ratio}-scaled sort, where an open category applies to some item \emph{by some amount more or less} (say \code{2.4} units) than to another item.
Notice that -- counterintuitively -- \emph{categorically}-scaled open sorts are not allowed.
}

If columns are named, they must be the same as the names in \code{descriptions}.
Either way, \code{assignments} and \code{descriptions} are always \emph{matched by index only}: the first column from \code{assignments}, must be the first element of \code{description}, and so forth.}

\item{descriptions}{a character vector giving the open-ended category description provided by the participant.
Can be named.
Defaults to \code{NULL}, in which case the user-defined categories are unknown (not recommended).}

\item{x}{a \link{psOpenSort}, created by \code{\link[=psOpenSort]{psOpenSort()}}.}

\item{object}{a \link{psOpenSort}, created by \code{\link[=psOpenSort]{psOpenSort()}}.}

\item{edge_codings}{a tibble with category description indeces in the first column, and arbitrary metadata \emph{about the descriptions} in limited later columns:
\itemize{
\item 2nd column will be mapped to line color,
\item 3rd column will be mapped to line type (and must be discrete).
Later columns will be ignored.
Useful if participants or researchers have coded the open-ended descriptions in some way.
Category description indeces must be a subset of the column names in \code{x}.
If more then one code applies to a category, multiple rows with identical categories can exist, and multiple, "fanned-out" edges will be drawn between the respective nodes.
See note.
Defaults to \code{NULL}.
}}

\item{str_wrap_width}{integer scalar, giving the maximum number of characters after which to insert a newline, passed on to \code{\link[stringr:str_wrap]{stringr::str_wrap()}}.
Defaults to \code{30}.
Useful for long descriptions.}

\item{...}{further arguments passed to methods.}
}
\value{
Object of class \code{psOpenSorts}.
}
\description{
Constructs S3 class for pensieve.
}
\details{
Open sorting categorizations \emph{cannot} be compared between participants, because each participants defines her own dimensions.

\strong{The canonical representation of open sorting data} is therefore a \emph{list} of matrices, one for each participant.
Every \emph{individual} matrix is a \code{\link[=psOpenSort]{psOpenSort()}} object, and together, they form a \code{\link[=psOpenSorts]{psOpenSorts()}} list.
The rows in these matrices are the items, the columns are the dimensions, and cells are the assignment.
Optional dimension descriptions are included as attributes of the matrices.

The canonical representation of open sorts in \code{\link[=psOpenSorts]{psOpenSorts()}} can be cumbersome to enter manually.
For \emph{logical} (nominally-scaled) open sorts, a simpler, but messier format can be conveniently entered as two separate spreadsheets of \code{descriptions_messy} and \code{assignments_messy} using \code{\link[=import_psOpenSorts]{import_psOpenSorts()}}.
}
\section{Functions}{
\itemize{
\item \code{import_psOpenSorts}: descriptions and \emph{logical} assignments from convenient, but messy format

\item \code{psOpenSort}: creates \emph{individual} open sort.
}}

\note{
When category is assigned, but never described, it is \code{TRUE} in the respective logical matrix entries and their description is \code{NA}:
This is still considered valuable, if incomplete information.
When a category is described, but never assigned, it is omitted from the data entirely.

When \emph{no} category was assigned to some item in \code{assignments_messay}, an empty character string \code{""} should be in the respective cell.

An \code{NA} value implies that the given participant never considered the given items \emph{at all}, across \emph{all} her categories.
Notice this implies \emph{limited scenarios of \code{NA}} for data entered in this messy, convenient form.
The more complicated cases, where a participant did consider \emph{some}, but \emph{not all} items in the assignment of a category, or -- equivalently -- all categories in their assessment of all items, cannot be recorded in this convenience format.
Such more granular \code{NA} records can, however, be recorded in the canonical data representation, where the respective cell of the items x category logical matrix would be \code{NA}.
If your data gathering procedure produces such granular \code{NA} records, do not use this convenience function.

To render the resulting \code{ggplot} object, you must \emph{manually} call \code{library(ggraph)} somewhere in your script (as per \href{https://github.com/thomasp85/ggraph/issues/85}{this limitation}).

If \code{codings} are added, the aesthetics are set \emph{for each individual \link{psOpenSort} separately}, which may make it hard to compare plots across participants.
To get consistent code aesthetics, consider applying \code{\link[ggraph:scale_edge_color_manual]{ggraph::scale_edge_color_manual()}} and friends.
}
\examples{
# create single open sort ====

# Lisas open sort, matching by index
assignments <- matrix(data = c(TRUE, FALSE, FALSE, TRUE),
                      nrow = 2,
                      dimnames = list(items = c("cat", "dog")))
descriptions <- c("a pet which largely takes care of itself",
                  NA)
lisa <- psOpenSort(assignments = assignments, descriptions = descriptions)

# Peters open sort, matching by name
assignments <- matrix(data = c(TRUE, TRUE, TRUE, FALSE, FALSE, FALSE),
                      nrow = 2,
                      dimnames = list(items = c("cat", "dog"),
                                      categories = c("in_homes",
                                                     "quiet",
                                                     "herbivore")))
descriptions <- c(in_homes = "Animal found in peoples homes.",
                  quiet = "Does not make a lot of noise.",
                  herbivore = "Eats plants.")
peter <- psOpenSort(assignments = assignments, descriptions = descriptions)

# Rebeccas open sort, without any descriptions provided
assignments <- matrix(data = c(FALSE, FALSE, TRUE, TRUE),
                      nrow = 2,
                      dimnames = list(handles = c("cat", "dog")))
rebecca <- psOpenSort(assignments = assignments, descriptions = NULL)


# now let's combine the individual sorts into a list ====
open_sorts <- psOpenSorts(open_sorts = list(lisa = lisa, peter = peter, rebecca = rebecca))


# create psOpenSorts from convenient input ====

ass <- matrix(data = c("A, B",
                       # meaning A and B are assigned
                       "",
                       # meaning no category assigned
                       "B",
                       # only B assigned
                       NA),
                       # item never considered for assignment across *all* categories or vice versa
                       nrow = 2,
                       ncol = 2,
                       dimnames = list(items = c("cat", "dog"),
                       people = c("tony", "amy")))
desc <- matrix(data = c("",
                        # will be treated as NA
                        NA,
                        # participant provided no description, but assigned the category
                        "lives in cage",
                        # described, but never assigned
                        NA,
                        # never assigned, never described will be removed
                        "actually a predator!",
                        "lives on a farm"
                        # described, but never assigned
                        ),
                        nrow = 3,
                        dimnames = list(categories = c("A", "B", "C"),
                        people = c("tony", "amy")))
# notice how individual *nominal* categories are pasted together in cells here;
# this convenient form *only* works for logical data
osorts_example <- import_psOpenSorts(assignments_messy = ass, descriptions_messy = desc)
ggplot2::autoplot(object = lisa)
ggplot2::autoplot(object = rebecca)

# no with codes
petercodes <- tibble::tibble(category = c("in_homes", "in_homes", "quiet", "herbivore"),
                             reference = c("location", "human interaction", NA, "animal diet"),
                             length = c("medium", "medium", "medium", "short")
                             # notice the duplicates to allow for multiple codes
                             )
library(ggraph)  # must be attached while running below
ggplot2::autoplot(object = peter, edge_codings = petercodes)
summary(peter)

ggplot2::autoplot(object = osorts_example)

}
\seealso{
Other S3 classes from `pensieve`: \code{\link{QItemStrata}},
  \code{\link{QPreSorts}}, \code{\link{QSorts}},
  \code{\link{correlate}}, \code{\link{extract}},
  \code{\link{make_grid}}, \code{\link{psConcourse}},
  \code{\link{psItems}}, \code{\link{psPeople}},
  \code{\link{score}}
}
\concept{S3 classes from `pensieve`}
