% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QItems.R
\name{QConcourse}
\alias{QConcourse}
\title{Check and make QConcourse}
\usage{
QConcourse(items, validate = TRUE)
}
\arguments{
\item{items}{A character matrix of full items, with named rows as item handles and named columns as languages.
Cells can be \code{NA} when full items are not available.
Full items must be unique by columns.}

\item{validate}{A logical flag, indicating whether the object will be validated.
Defaults to `TRUE`.}
}
\description{
Checks and makes QConcourse
}
\seealso{
Other validation helpers: \code{\link{QItemStrata}},
  \code{\link{QPeopleFeatures}}, \code{\link{QPreSorts}},
  \code{\link{QSorts}}

Other import helpers: \code{\link{QItemStrata}},
  \code{\link{QPeopleFeatures}}, \code{\link{QPreSorts}},
  \code{\link{QSorts}}, \code{\link{anonymize}},
  \code{\link{check.QPreSorts}}
}
