% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{QConcourse}
\alias{QConcourse}
\title{Check and make QConcourse}
\usage{
QConcourse(concourse, validate = TRUE)
}
\arguments{
\item{concourse}{A character matrix of full items, with named rows as item handles and named columns as languages.
Cells can be \code{NA} when full items are not available.
Full items must be unique by columns.}

\item{validate}{A logical flag, indicating whether the object will be validated on construction.
Defaults to `TRUE`.}
}
\description{
Check and make QConcourse

Construct S3 class.
}
\seealso{
Other S3 classes from `pensieve`: \code{\link{QItemStrata}},
  \code{\link{QPeopleFeatures}}, \code{\link{QPreSorts}},
  \code{\link{QSorts}}, \code{\link{correlate}},
  \code{\link{extract}}, \code{\link{score}}

Other import helpers: \code{\link{QItemStrata}},
  \code{\link{QPeopleFeatures}}, \code{\link{QPreSorts}},
  \code{\link{QSorts}}, \code{\link{anonymize}},
  \code{\link{check}}
}
