% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{QItemConcourse}
\alias{QItemConcourse}
\title{Check and make QItemConcourse}
\usage{
QItemConcourse(concourse, validate = TRUE)
}
\arguments{
\item{concourse}{Character matrix of \strong{full items}, with named rows as item handles and named columns as languages.
Cells can be \code{NA} when full items are not available.
Full items must be unique by columns.}

\item{validate}{A logical flag, indicating whether the object will be validated on construction.
Defaults to `TRUE`.}
}
\description{
Check and make QItemConcourse

Construct S3 class.
}
\seealso{
Other S3 classes from `pensieve`: \code{\link{QItemStrata}},
  \code{\link{QPeopleFeatures}}, \code{\link{QPreSorts}},
  \code{\link{QSorts}}, \code{\link{correlate}},
  \code{\link{extract}}, \code{\link{make_grid}},
  \code{\link{score}}
}
