% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qItems.R
\name{qItems}
\alias{check_qItems}
\alias{qItems}
\title{Check and make qItems.}
\usage{
qItems(q_set, concourse = NULL, lookup = NULL)

check_qItems(q_set, concourse = NULL, lookup = NULL)
}
\arguments{
\item{q_set}{A character vector of item handles.
Handles must be unique and valid R names.
Handles must be a subset of \code{concourse}.}

\item{concourse}{A character matrix of full items, with named rows as item handles and named columns as languages.
Defaults to \code{NULL}, in which case full items are not provided.
Full items must be unique by columns.}

\item{lookup}{A data frame or tibble of manually defined item IDs with the first column as item handles.
Can have several columns of alternative item IDs.
Item IDs must be unique by columns.
All item handles from \code{q_set} must be defined.
Defaults to \code{NULL}, in which case items are automatically identified (recommended).}
}
\value{
A character vector of item handles, all as unique and valid R names.
}
\description{
Checks and makes qItems from concourse, q-set and optional lookup table.
}
\details{
The basis of a Q study is a concourse of possible items, from which \emph{some} items are sampled into the Q-Set.
Q-Set and concourse are sometimes accompanied by manual lookup table, though automatic procedures are recommended.
These three objects are here tested for internal consistency, and combined into the canonical \code{qItems} S3 object.
}
\seealso{
Other import: \code{\link{count_cooccur}},
  \code{\link{import_qcat}}
}

