% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qItems.R
\name{QItems}
\alias{QItems}
\title{Check and make QItems.}
\usage{
QItems(items, validate = TRUE)
}
\arguments{
\item{items}{A character matrix of full items, with named rows as item handles and named columns as languages.
Cells can be \code{NA} when full items are not available.
Full items must be unique by columns.}

\item{validate}{A logical flag, indicating whether the object will be validated.
Defaults to \code{TRUE}.}
}
\description{
Checks and makes QItems.
}
\seealso{
Other import helpers: \code{\link{QPeopleFeatures}},
  \code{\link{QPreSorts}}, \code{\link{QSorts}},
  \code{\link{anonymize}}, \code{\link{check.QPreSorts}}

Other validation helpers: \code{\link{QPeopleFeatures}},
  \code{\link{QPreSorts}}, \code{\link{QSorts}}
}

