% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{QItemStrata}
\alias{QItemStrata}
\title{Check and make QItemStrata}
\usage{
QItemStrata(strata, validate = TRUE)
}
\arguments{
\item{strata}{A logical array of arbitrary dimensions, with first dimension (rows) as item handles, and higher dimensions as orthogonal sampling strata.}

\item{validate}{a logical flag, indicating whether the object will be validated on construction.
Defaults to `TRUE`.}
}
\description{
Checks and makes QItemStrata, the item sampling structure

Construct S3 class for pensieve.
}
\seealso{
Other S3 classes from `pensieve`: \code{\link{QPreSorts}},
  \code{\link{QSorts}}, \code{\link{correlate}},
  \code{\link{extract}}, \code{\link{make_grid}},
  \code{\link{pensieveAssignments}},
  \code{\link{pensieveConcourse}},
  \code{\link{pensieveItems}},
  \code{\link{pensievePeople}}, \code{\link{score}}

Other import helpers: \code{\link{QPreSorts}},
  \code{\link{QSorts}}, \code{\link{anonymize}}
}
