% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorts.R
\name{QPeopleFeatures}
\alias{QPeopleFeatures}
\title{Check and make QPeopleFeatures}
\usage{
QPeopleFeatures(p_feat, validate = TRUE)
}
\arguments{
\item{p_feat}{A tibble, with one row per participant.
First column must be the participant names, same as the rownames from \code{\link{QSorts}}.}

\item{validate}{A logical flag, indicating whether the object will be validated on construction.
Defaults to `TRUE`.}
}
\description{
Checks and makes QPeopleFeatures, a tibble with arbitrary additional information on the participating people-variables.

Construct S3 class.
}
\seealso{
Other S3 classes from `pensieve`: \code{\link{QItemConcourse}},
  \code{\link{QItemStrata}}, \code{\link{QPreSorts}},
  \code{\link{QSorts}}, \code{\link{correlate}},
  \code{\link{extract}}, \code{\link{make_grid}},
  \code{\link{score}}

Other import helpers: \code{\link{QItemStrata}},
  \code{\link{QPreSorts}}, \code{\link{QSorts}},
  \code{\link{anonymize}}, \code{\link{check}}
}
