% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QData.R
\name{QPreSorts}
\alias{QPreSorts}
\title{Check and make QPreSorts}
\usage{
QPreSorts(presorts, validate = TRUE)
}
\arguments{
\item{presorts}{An integer matrix, with named rows as item handles, named columns as participant names and cells as presorts.
\code{-1L} for `negative`, \code{0L} for `neutral` and \code{1L} for `positive`.}

\item{validate}{A logical flag, indicating whether the object will be validated.
Defaults to \code{TRUE}.}
}
\description{
Checks and makes QPreSorts
}
\note{
\code{presorts} are stored as \code{integer()} because R does not allow factor matrices.
Pre-sorting piles are, of course, \emph{categorical} information and should be treated as such.
}
\seealso{
Other import helpers: \code{\link{QConcourse}},
  \code{\link{QPeopleFeatures}}, \code{\link{QSorts}},
  \code{\link{anonymize}}, \code{\link{check.QPreSorts}}

Other validation helpers: \code{\link{QConcourse}},
  \code{\link{QPeopleFeatures}}, \code{\link{QSorts}}
}
