% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sorts.R
\name{QPreSorts}
\alias{QPreSorts}
\title{Check and make QPreSorts}
\usage{
QPreSorts(presorts, validate = TRUE)
}
\arguments{
\item{presorts}{An integer matrix, with named rows as item handles, named columns as participant names and cells as presorts.
\code{-1L} for \code{negative}, \code{0L} for \code{neutral} and \code{1L} for \code{positive}.}

\item{validate}{A logical flag, indicating whether the object will be validated.
Defaults to `TRUE`.}
}
\description{
Checks and makes QPreSorts
}
\note{
\code{presorts} are stored as \code{integer()} because R does not allow factor matrices.
Pre-sorting piles are, of course, \code{categorical} information and should be treated as such.
}
\seealso{
Other validation helpers: \code{\link{QConcourse}},
  \code{\link{QCors}}, \code{\link{QItemStrata}},
  \code{\link{QPeopleFeatures}}, \code{\link{QSorts}}

Other import helpers: \code{\link{QConcourse}},
  \code{\link{QItemStrata}}, \code{\link{QPeopleFeatures}},
  \code{\link{QSorts}}, \code{\link{anonymize}},
  \code{\link{check}}
}
