% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QData.R
\name{QSorts}
\alias{QSorts}
\title{Check and make QSorts}
\usage{
QSorts(qsorts, validate = TRUE)
}
\arguments{
\item{qsorts}{An integer array with item handles as first dimension, people as second dimension, arbitrary dimensions thereafter, and item positions in cells.
Dimensions must be named.}

\item{validate}{A logical flag, indicating whether the object will be validated.
Defaults to \code{TRUE}.}
}
\description{
Checks and makes QSorts
}
\seealso{
Other import helpers: \code{\link{QItems}},
  \code{\link{QPeopleFeatures}}, \code{\link{QPreSorts}},
  \code{\link{anonymize}}, \code{\link{check.QPreSorts}}

Other validation helpers: \code{\link{QItems}},
  \code{\link{QPeopleFeatures}}, \code{\link{QPreSorts}}
}

