% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psItemContent.R
\name{render_items}
\alias{render_items}
\title{Render text items.}
\usage{
render_items(items, lang = NULL, tex = NULL, pdf = NULL, svg = NULL,
  grob = NULL, ...)
}
\arguments{
\item{items}{\code{[character()]} giving the \emph{participant-facing} \strong{item content}.
Can be named to provide short, \emph{researcher-facing} \strong{item handles}.}

\item{lang}{\code{[character(1)]} giving a language code for \emph{all} items, such as \code{en_US}.
Used for multilingual typsetting support via \href{https://ctan.org/pkg/babel}{LaTeX's babel package}.
Must be one of:
\itemize{
\item \code{NULL} (default), in which case there is no multilingual typesetting support.
\item a \href{https://tools.ietf.org/html/bcp47}{valid BCP 47 language code} supported by pandoc.
See \code{pensieve:::format_latex$avail_opts$langs} for all available languages.
Ignored unless \code{type = "text"}.
}}

\item{tex}{\code{[list(character())]} giving a list of manually produced LaTeX markup, one for each \code{items}.
Defaults to \code{NULL}, in which case the LaTeX markup is rendered automatically (recommended).}

\item{pdf}{\code{[list(raw())]} giving a list of manually produced PDFs, one for each \code{items}.
Defaults to \code{NULL}, in which case the PDF is rendered automatically (recommended).}

\item{svg}{\code{[list(raw())]} giving a list of manually produced SVGs, one for each \code{items}.
Defaults to \code{NULL}, in which case the SVG is rendered automatically (recommended).}

\item{grob}{\code{[list(grob())]} giving a list of manually produced grobs, one for each \code{items}.
Defaults to \code{NULL}, in which case the grob is rendered automatically (recommended).}

\item{...}{further arguments to be passed to methods.}
}
\description{
Renders character vectors of items to pdf, svg and R graphics (grid).
}
\details{
It is often helpful to have an authoritative, typeset version of text items, ready for for printing, web publishing or interpretation.
Rendered text items should meet several criteria:
\itemize{
\item They should always fit on the \emph{same} card size, both for practical reasons and to emphasize the equal "significance" of items, even if they are of different length.
Card sizes can be arbitrary, specified via \code{paperwidth},\code{paperheight}.
\item They should always \emph{look identical}, no matter where a participant or researcher encounters them.
Even slight variations in, for example, line wrapping, might give an item a slightly different emphasis.
\item Given the central status of text items in the methodology and the package, they should by \emph{typeset professionally}.
}

To meet these criteria, text items are
\enumerate{
\item converted to \strong{LaTeX}, using \href{https://pandoc.org}{pandoc}, then
\item compiled to \strong{PDF}, using \href{https://www.latex-project.org}{LaTeX} via \link[tools:texi2pdf]{tools::texi2pdf}, then
\item converted to \strong{SVG}, using \href{https://github.com/dawbarton/pdf2svg}{pdf2svg}, then
\item imported to \strong{R Graphics} (grid graphics, to be precise) via \code{\link[grImport2:readPicture]{grImport2::readPicture()}}.
Items are now fully available to the R Graphics system and can be used wherever \code{\link[graphics:plot]{graphics::plot()}} (or, to be precise, \code{\link[grid:grid.draw]{grid::grid.draw()}}) works.
}

At each step of this necessary, but rather long conversion pipeline more (system) dependencies are required and asserted.

Additionally, because some of the intermediary formats cannot be easily or fully converted, downstream outputs may be faulty.
Always use the \emph{earliest} possible output from the above conversion pipeline to maximize fidelity to the original PDF.

When \code{items} are named with item handles, such handles are used as file names.

Because items always have to fit on one page, this function errors out when the rendered item would fill more than one page.
}
