% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psItemContent.R
\name{render_items}
\alias{render_items}
\title{Render text items to pdf and svg.}
\usage{
render_items(items, lang, output_dir = NULL, fontsize = NULL,
  paperwidth = 8.5, paperheight = 5.4, top = 0.5, bottom = 0.5,
  left = 0.5, right = 0.5, units = "cm", alignment = "left")
}
\arguments{
\item{items}{\code{[character()]} giving the \emph{participant-facing} \strong{item content}.
Can be named to provide a short, \emph{researcher-facing} \strong{item handles}.}

\item{lang}{\code{[character(1)]} giving a language code for \emph{all} items, such as \code{en-US}.
Used for multilingual typsetting support via \href{https://ctan.org/pkg/babel}{LaTeX's babel package}.
Must be one of:
\itemize{
\item \code{NULL} (default), in which case there is no multilingual typesetting support.
\item a \href{https://tools.ietf.org/html/bcp47}{valid BCP 47 language code}.
Ignored unless \code{type = "text"}.
}}

\item{output_dir}{\code{[character(1)]}
giving directory relative from working directory root \code{\link[base:getwd]{base::getwd()}}.
Best constructed with \code{\link[base:file.path]{base::file.path()}}.
Defaults to \code{NULL}, in which case items are rendered to the working directory root.}

\item{fontsize}{\code{[character(1)]}
giving a \href{https://en.wikibooks.org/wiki/LaTeX/Fonts#Sizing_text}{LaTeX fontsize} for \emph{all} items.
See \code{pensieve:::latex$options$fontsize} for valid options.
Defaults to \code{NULL}, in which case the maximum possible fontsize is sought and chosen, by which \emph{all} items still fit on one page.}

\item{paperwidth}{\code{[numeric(1)]} giving the width of cards in \code{units}, passed on to \href{https://ctan.org/pkg/geometry}{LaTeX geometry package}.
For good typographical results, should be as close as possible to the \emph{actual} physical measurements of cards encountered by users.
Defaults to \code{8.5}.}

\item{paperheight}{\code{[numeric(1)]} giving the height of cards in \code{units}, passed on to \href{https://ctan.org/pkg/geometry}{LaTeX geometry package}.
For good typographical results, should be as close as possible to the \emph{actual} physical measurements of cards encountered by users.
Defaults to \code{5.4}.}

\item{top}{\code{[numeric(1)]} giving the margin in \code{units}, passed on to \href{https://ctan.org/pkg/geometry}{LaTeX geometry package}.
Defaults to \code{0.5}.}

\item{bottom}{\code{[numeric(1)]} giving the margin in \code{units}, passed on to \href{https://ctan.org/pkg/geometry}{LaTeX geometry package}.
Defaults to \code{0.5}.}

\item{left}{\code{[numeric(1)]} giving the margin in \code{units}, passed on to \href{https://ctan.org/pkg/geometry}{LaTeX geometry package}.
Defaults to \code{0.5}.}

\item{right}{\code{[numeric(1)]} giving the margin in \code{units}, passed on to \href{https://ctan.org/pkg/geometry}{LaTeX geometry package}.
Defaults to \code{0.5}.}

\item{units}{\code{[character(1)]} giving the units for the above dimensions, must be one of:
\itemize{
\item "cm" for metric system,
\item "in" for inches.
Defaults to \code{"cm"}.
}}

\item{alignment}{\code{[character(1)]} giving the alignment for the text, must be one of:
\itemize{
\item "justified",
\item "left",
\item "right" or
\item "center".
Defaults to \code{"left"}.
}}
}
\description{
Renders character vectors of items to a pdf, using pandoc and latex, then converts pdf to svg using pdf2svg.
}
\details{
It is often helpful to have an authoritative, and professionally typeset version of items.
This function produces renders items.
When \code{items} are named with item handles, such handles are used as file names.
Because items always have to fit on one page, this function errors out when the rendered item would fill more than one page.
}
