% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\alias{QScores}
\alias{check.QScores}
\title{Factor scores.}
\usage{
score(loas, csorts)

QScores(scores, validate = TRUE)

\method{check}{QScores}(x)
}
\arguments{
\item{loas}{A numerical matrix with loadings.}

\item{csorts}{\code{[array()]}
An numeric array with people as rows, item handles as columns, arbitrary dimensions thereafter (such as for multiple sorting dimensions), and item positions in cells.}

\item{scores}{A numerical matrix with factor scores.}

\item{validate}{a logical flag, indicating whether the object will be validated on construction.
Defaults to \code{TRUE}.}

\item{x}{class object created by respective constructor function.}
}
\description{
Calculates the factor scores as loadings-weighted averages of raw Q-Sorts (regression scores).

Construct S3 class for pensieve.

Check S3 class.
}
\section{Methods (by generic)}{
\itemize{
\item \code{check}: validation
}}

\examples{
csorts <- civicon_2014$qData$sorts[,,"before"]  # preparatory step
cors <- correlate(csorts = csorts)  # preparatory step
loas <- extract(cors = cors, nfactors = 3, fa_type = "pca")  # preparatory step
scores <- score(loas = loas, csorts = csorts)
# this just assigns the class, without validation (not recommended)
scores <- QScores(scores = scores, validate = FALSE)
# this validates the class
check(x = scores)
}
\seealso{
Other analysis functions: \code{\link{correlate}},
  \code{\link{extract}}

Other S3 classes from `pensieve`: \code{\link{correlate}},
  \code{\link{extract}}, \code{\link{psClosedSorts}},
  \code{\link{psGrid}}, \code{\link{psItemContent}},
  \code{\link{psOpenSorts}}, \code{\link{psOpenSort}},
  \code{\link{psPeople}}
}
\concept{S3 classes from `pensieve`}
\concept{analysis functions}
\concept{scoring functions}
