% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score}
\alias{score}
\alias{QScores}
\alias{check.QScores}
\title{Factor scores.}
\usage{
score(loas, sorts)

QScores(scores, validate = TRUE)

\method{check}{QScores}(x)
}
\arguments{
\item{loas}{A numerical matrix with loadings.}

\item{sorts}{An integer array with item handles as first dimension, people as second dimension, arbitrary dimensions thereafter, and item positions in cells.
Dimensions must be named.}

\item{scores}{A numerical matrix with factor scores.}

\item{validate}{a logical flag, indicating whether the object will be validated on construction.
Defaults to `TRUE`.}

\item{x}{class object created by respective constructor function.}
}
\description{
Calculates the factor scores as loadings-weighted averages of raw Q-Sorts (regression scores).

Construct S3 class for pensieve.

Check S3 class.
}
\details{
`check()`, `test()`, `assert()` and `expect()` are extensions the family of functions from the [checkmate](https://github.com/mllg/checkmate) package.
`check()` returns `TRUE` or the error message, `assert()` only returns a message in case of an error, `test()` returns `TRUE` or `FALSE`.
`expect()` is for internal use with testing via [`testthat`](https://github.com/hadley/testthat)).
`need()` returns `NULL` or the error message and is for interal use with the accio web frontend inside [shiny::validate()].
}
\section{Methods (by generic)}{
\itemize{
\item \code{check}: validation
}}

\examples{
sorts <- civicon_2014$qData$sorts[,,"before"]  # preparatory step
cors <- correlate(sorts = sorts)  # preparatory step
loas <- extract(cors = cors, nfactors = 3, fa_type = "pca")  # preparatory step
scores <- score(loas = loas, sorts = sorts)
# this just assigns the class, without validation (not recommended)
scores <- QScores(scores = scores, validate = FALSE)
# check object and friends ...
# check(x)  # returns TRUE or error message
# test(x)  # returns TRUE or FALSE
# assert(x)  # returns error or silently object
# need(x)  # returns NULL or error message
# this validates the class
check(x = scores)
}
\seealso{
Other analysis functions: \code{\link{correlate}},
  \code{\link{extract}}

Other S3 classes from `pensieve`: \code{\link{QItemStrata}},
  \code{\link{QPreSorts}}, \code{\link{QSorts}},
  \code{\link{correlate}}, \code{\link{extract}},
  \code{\link{make_grid}}, \code{\link{psConcourse}},
  \code{\link{psItems}}, \code{\link{psOpenSorts}},
  \code{\link{psPeople}}

Other validation functions: \code{\link{QSorts}},
  \code{\link{check}}, \code{\link{correlate}},
  \code{\link{extract}}, \code{\link{make_grid}}
}
