% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psItemContent.R
\name{wrap_in_latex_env}
\alias{wrap_in_latex_env}
\alias{wrap_in_latex_fontsize}
\alias{wrap_in_latex_alignment}
\title{Wrap latex string in latex environment}
\usage{
wrap_in_latex_env(env, tex)

wrap_in_latex_fontsize(fontsize_local = "tiny", tex)

wrap_in_latex_alignment(alignment = "justified", tex)
}
\arguments{
\item{env}{\code{[character(1)]} giving a latex environment.}

\item{tex}{\code{[character(1)]} giving some latex string.}

\item{fontsize_local}{\code{[character(1)]} giving a valid \href{https://en.wikibooks.org/wiki/LaTeX/Fonts#Sizing_text}{LaTeX font size}.

Must be one of:
\itemize{
\item \code{'tiny'} (\strong{default})
\item \code{'scriptsize'}
\item \code{'footnotesize'}
\item \code{'small'}
\item \code{'normalsize'}
\item \code{'large'}
\item \code{'Large'}
\item \code{'LARGE'}
\item \code{'huge'} or
\item \code{'Huge'}
}}

\item{alignment}{\code{[character(1)]} giving the alignment of the text.

Must be one of:
\itemize{
\item \code{'justified'} (\strong{default})
\item \code{'left'}
\item \code{'right'} or
\item \code{'center'}
}}
}
\value{
\code{[character(1)]} a latex string
}
\description{
These are helper functions to apply latex environments.
}
\section{Functions}{
\itemize{
\item \code{wrap_in_latex_fontsize}: Apply local fontsize

\item \code{wrap_in_latex_alignment}: Apply alignment
}}

\keyword{internal}
